% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{add_rank_and_group}
\alias{add_rank_and_group}
\title{add_rank_and_group}
\usage{
add_rank_and_group(TSF, N = 5, sectorNe = NULL, untie = 1.5)
}
\description{
add the rank and groups by factorscores
}
