% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{backtestPar.default}
\alias{backtestPar.default}
\title{backtestPar.default}
\usage{
backtestPar.default()
}
\value{
a \bold{backtestPar} object with default parametres
}
\description{
set the backtesting with default parametres
}
\seealso{
Other backtestPar setting functions: \code{\link{backtestPar.IC}},
  \code{\link{backtestPar.Ngroup}},
  \code{\link{backtestPar.fee}},
  \code{\link{backtestPar.longshort}}
}
\author{
Ruifei.Yin
}
