% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{backtestPar.fee}
\alias{backtestPar.fee}
\alias{setbacktestPar.fee}
\title{backtestPar.fee}
\usage{
backtestPar.fee(backtestPar = backtestPar.default(), secu = 0, future = 0)

setbacktestPar.fee(backtestPar, secu, future)
}
\arguments{
\item{backtestPar}{a \bold{backtestPar} object}

\item{secu}{a numeric, giving buy and sell (one side) fee of the security}

\item{future}{a numeric, giving  buy and sell (one side) fee of the future}
}
\value{
a \bold{backtestPar} object
}
\description{
set the fee related parametres of the backtesting
}
\author{
Ruifei.Yin
}
\seealso{
\code{\link{port.backtest}},\code{\link{getrtn.LSH}},\code{\link{getrtn.LBH}}

Other backtestPar setting functions: \code{\link{backtestPar.IC}},
  \code{\link{backtestPar.Ngroup}},
  \code{\link{backtestPar.default}},
  \code{\link{backtestPar.longshort}}
}

