% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{backtestPar.IC}
\alias{backtestPar.IC}
\alias{setbacktestPar.IC}
\title{backtestPar.IC}
\usage{
backtestPar.IC(backtestPar = backtestPar.default(), stat = c("pearson",
  "spearman"))

setbacktestPar.IC(backtestPar, stat)
}
\arguments{
\item{backtestPar}{a \bold{backtestPar} object}

\item{stat}{character string,indicating the methods to compute IC,could be "pearson" or "spearman".}
}
\value{
a \bold{backtestPar} object
}
\description{
set the IC related parametres of the backtesting
}
\author{
Ruifei.Yin
}
\seealso{
\code{\link{seri.IC}}

Other backtestPar setting functions: \code{\link{backtestPar.Ngroup}},
  \code{\link{backtestPar.default}},
  \code{\link{backtestPar.fee}},
  \code{\link{backtestPar.longshort}}
}

