% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{backtestPar.longshort}
\alias{backtestPar.longshort}
\alias{setbacktestPar.longshort}
\title{backtestPar.longshort}
\usage{
backtestPar.longshort(backtestPar = backtestPar.default(), topN = 50,
  topQ = NA, factorNA = c("na", "mean", "median", "min", "max", "sectmean"),
  pick.sectorNe = FALSE, sectorAttr = defaultSectorAttr(), force_in = 0,
  buffer_keep = 0, buffer_rate = 0, init_port = NULL,
  holdingEndT = Sys.Date(), wgtType = c("eq", "fv", "fvsqrt", "custom",
  "ffv", "ffvsqrt"), wgt.sectorNe = FALSE, wgt.max = NA, bmk = "EI000300",
  hedge.rebFreq = "month", hedge.posi = 1, hitFreq = "month")

setbacktestPar.longshort(backtestPar, topN, topQ, factorNA, pick.sectorNe,
  sectorAttr, force_in, buffer_keep, buffer_rate, init_port, holdingEndT,
  wgtType, wgt.sectorNe, wgt.max, bmk, hedge.rebFreq, hedge.posi, hitFreq)
}
\arguments{
\item{backtestPar}{a \bold{backtestPar} object}

\item{topN}{an integer,giving the numbers of the assets to be selected into the portfolio.}

\item{topQ}{a numeric,giving the percentage of the assets to be selected into the portfolio.}

\item{holdingEndT}{the ending date of the holding portfolio.}

\item{wgtType}{a character string, giving the weighting type of portfolio,which could be "eq"(equal),"fv"(floatValue),"fvsqrt"(sqrt of floatValue) or "custom".}

\item{wgt.sectorNe}{a logic. If true, the wgt will be neutralized by sector.}

\item{wgt.max}{a integer, giving the max weight limit which could be set on a single stock. If NA(the default value), with no limit. See \code{\link{port.substitute}}.}

\item{bmk}{a character string,giving the stockID of the benchmark index, eg. "EI000300".}

\item{hedge.rebFreq}{giving the rebalance freq when computing the hedged rtn.}

\item{hitFreq}{indicating the interval when computing the hitRatio of rtn. An interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s".See \code{\link{cut.Date}} for detail.}

\item{hedege.posi}{a numeric, giving the position of the hedging assets}
}
\value{
a \bold{backtestPar} object
}
\description{
set the long-short related parametres of the backtesting
}
\author{
Ruifei.Yin
}
\seealso{
\code{\link{getPort}}, \code{\link{addwgt2port}}, \code{\link{port.substitute}}, \code{\link{port.backtest}}, \code{\link{getrtn.LSH}}, \code{\link{getrtn.LBH}}, \code{\link{tables.longshort}}

Other backtestPar setting functions: \code{\link{backtestPar.IC}},
  \code{\link{backtestPar.Ngroup}},
  \code{\link{backtestPar.default}},
  \code{\link{backtestPar.fee}}
}

