% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{build.lcdb.RegTables}
\alias{build.lcdb.RegTables}
\title{build local database's regression tables}
\usage{
build.lcdb.RegTables(FactorLists)
}
\description{
build local database's regression tables
}
\examples{
build.lcdb.RegTables(FactorLists)
}

