% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{buildBoxConstr}
\alias{buildBoxConstr}
\title{buildBoxConstr}
\usage{
buildBoxConstr(...)
}
\description{
build box constraint in optimization function
}
\examples{
boxConstr <- buildBoxConstr(EI000905=c(0.75,0.85))
boxConstr <- buildBoxConstr(EI000300=c(0.75,0.85))
}

