% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{buildFactorExp}
\alias{buildFactorExp}
\title{buildFactorExp}
\usage{
buildFactorExp(sectorall = c(-0.05, 0.05), ...)
}
\description{
build factor exposure setting in optimization function
}
\examples{
fexp <- buildFactorExp()
fexp <- buildFactorExp(disposition_=c(-0.01,1),beta_ = c(-0.01, 1))
fexp <- buildFactorExp(sectorall=c(-0.05,0.05),disposition_=c(-0.01,1),ES33480000=c(0.15,0.25))
}
