% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{buildFactorLists}
\alias{buildFactorLists}
\alias{buildFactorLists_lcfs}
\title{buildFactorLists}
\usage{
buildFactorLists(..., factorStd, sectorAttr, factorOutlier, factorNA)

buildFactorLists_lcfs(factorIDs, factorStd = "none",
  sectorAttr = defaultSectorAttr(), factorOutlier = 3, factorNA = "na")
}
\arguments{
\item{...}{one or more FactorList. See more detail in \code{\link{buildFactorList}}.}

\item{factorIDs}{a character vector of factorID, available in table \code{CT_FactorLists()} .}
}
\value{
a list of \bold{FactorList}
}
\description{
get a list of \bold{FactorList}, which is offen used in "multi-factor model builbing" or in "multifactor comparison".
}
\note{
In function \code{buildFactorLists}, the settings of factorStd, factorOutlier, factorNA of factors will be change samely by the given arguments, if they are not missing.

function \code{buildFactorLists_lcfs} give a simply and direct way to build a \bold{list of factorlist} through "lcfs", given a vector of factorID. Here, the settings of factorStd, factorOutlier, factorNA of factors will be set samely by the given arguments, and the other settings will be get through "lcfs" (\code{CT_FactorLists()}). See also \code{\link{buildFactorList_lcfs}}.
}
\examples{
FactorLists <- buildFactorLists(
  buildFactorList(factorFun="gf.pct_chg_per",
                  factorPar=list(N=60),
                  factorDir=-1),
  buildFactorList(factorFun="gf.float_cap",
                  factorPar=list(),
                  factorDir=-1),
  buildFactorList(factorFun="gf.PE_ttm",
                  factorPar=list(),
                  factorDir=-1)
)
# - change some public arguments.
FactorLists2 <- buildFactorLists(
  buildFactorList(factorFun="gf.pct_chg_per",
                  factorPar=list(N=60),
                  factorDir=-1),
  buildFactorList(factorFun="gf.float_cap",
                  factorPar=list(),
                  factorDir=-1),
  buildFactorList(factorFun="gf.PE_ttm",
                  factorPar=list(),
                  factorDir=-1),
  factorStd = "sectorNe",
  factorNA = "mean"
)
# - through lcfs
factorIDs <- c("F000001","F000002","F000005")
FactorLists3 <- buildFactorLists_lcfs(factorIDs)
}
