% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{buildWgtSet}
\alias{buildWgtSet}
\title{buildWgtSet}
\usage{
buildWgtSet(wgt = c(0, 0.01), ...)
}
\description{
build weight setting in optimization function
}
\examples{
wgt <- buildWgtSet()
wgt <- buildWgtSet(wgt=c(0,0.01),EQ601318=c(0.05,0.05))
}

