% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{calDelta}
\alias{calDelta}
\title{calDelta}
\usage{
calDelta(TS, dure, datasrc = c("local", "regResult"), reg_results,
  nwin = 250)
}
\arguments{
\item{dure}{a period object from package \code{lubridate}. (ie. \code{months(1),weeks(2)}. See example in \code{\link{trday.offset}}.) If null, then get periodrtn between \code{date} and the next \code{date}, else get periodrtn of '\code{dure}' starting from \code{date}.}

\item{reg_results}{is a factor return dataframe.}

\item{nwin}{is rolling window.}
}
\value{
delta.
}
\description{
calculate residual's Delta
}
\examples{
delta <- calDelta(TS,dure)
}
