% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{chart.FctRtn_scatter}
\alias{chart.FctRtn_scatter}
\title{chart.FctRtn_scatter}
\usage{
chart.FctRtn_scatter(TSFR, sample = NULL, ncol = NULL)
}
\description{
chart.FctRtn_scatter
}
\examples{
modelPar <- modelPar.default()
TSFR <- Model.TSFR(modelPar)
chart.FctRtn_scatter(TSFR,25)
}
