% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{chart.PA.attr}
\alias{chart.PA.attr}
\title{chart.PA.attr}
\usage{
chart.PA.attr(PA_tables, riskfnames, plotInd = FALSE, attributeAnn = TRUE)
}
\description{
chart.PA.attr
}
\examples{
riskfnames <- sapply(riskLists,'[[','factorName')
chart.PA.attr(PA_tables,riskfnames)
chart.PA.attr(PA_tables,riskfnames,plotInd=TRUE)
}

