% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{chart.PA.exposure}
\alias{chart.PA.exposure}
\title{chart.PA.exposure}
\usage{
chart.PA.exposure(PA_tables, riskfnames, plotInd = FALSE)
}
\description{
chart.PA.exposure
}
\examples{
riskfnames <- sapply(riskLists,'[[','factorName')
chart.PA.exposure(PA_tables,riskfnames,plotInd=FALSE)
chart.PA.exposure(PA_tables,riskfnames,plotInd=TRUE)
}

