% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{chart.reg.fRtnWealthIndex}
\alias{chart.reg.fRtnWealthIndex}
\title{chart.reg.fRtnWealthIndex}
\usage{
chart.reg.fRtnWealthIndex(reg_results, facet = FALSE)
}
\description{
chart.reg.fRtnWealthIndex
}

