% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{chart.TSFRScatter}
\alias{chart.TSFRScatter}
\title{chart.TSFRScatter}
\usage{
chart.TSFRScatter(TSFR, Nbin = "day", plotPar)
}
\arguments{
\item{TSFR}{a \bold{TSFR} object}

\item{Nbin}{the number of the groups the timespan is cut to when plotting the scatter by time series.It could also be a character of interval specification,See \code{\link{cut.Date}} for detail. The default value is "day",which means no cutting, the scatters of every date are ploted.}

\item{plotPar}{Optional.a \bold{plotPar} object,if not missing,then extract pars from plotPar}
}
\value{
a list, with 2 plot objects: 
   \itemize{
   \item scatter.overall: the all span scatter
   \item scatter.ts: the time series scatter,which is display in diffrent facets
   }
}
\description{
plot the \bold{TSFR} object with scatter,which demonstrating the relationship between the factorscore and periodrtn.
}
\examples{
modelPar <- modelPar.default()
TSFR <- Model.TSFR(modelPar)
chart.TSFRScatter(TSFR,25)
}
\author{
Ruifei.Yin
}

