% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{default.factorName}
\alias{default.factorName}
\title{default.factorName}
\usage{
default.factorName(factorFun, factorPar, factorDir)
}
\description{
default.factorName
}
\examples{
factorFun = "gf.pct_chg_per"
default.factorName(factorFun,"2,3")
default.factorName(factorFun,"2,3",-1)
default.factorName(factorFun,list(N=60,M=30))
default.factorName(factorFun,"20,\\"IF00\\"")
default.factorName(factorFun,"lag=20,id=\\"IF00\\"")
}

