% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{f_rtn_cov_delta}
\alias{f_rtn_cov_delta}
\alias{getfRtn}
\alias{getfCov}
\alias{getDelta}
\title{factor return,covariance and delta}
\usage{
f_rtn_cov_delta(dure = months(1), rolling = FALSE, rtntype = c("mean",
  "forcast"), covtype = c("shrink", "simple"), nwin = 24, reg_results)

getfRtn(dure = months(1), rolling = FALSE, rtntype = c("mean", "forcast"),
  nwin = 24, reg_results)

getfCov(dure = months(1), rolling = FALSE, covtype = c("shrink",
  "simple"), nwin = 24, reg_results)

getDelta(dure = months(1), rolling = FALSE, nwin = 24, reg_results)
}
\arguments{
\item{dure}{a period object from package \code{lubridate}. (ie. \code{months(1),weeks(2)}. See example in \code{\link{trday.offset}}.) If null, then get periodrtn between \code{date} and the next \code{date}, else get periodrtn of '\code{dure}' starting from \code{date}.}

\item{rolling}{default value is \code{FALSE}, if value is \code{TRUE} means the data period is \code{nwin} forward.}

\item{rtntype}{is method to caculate factor return,\bold{mean} means average of historical data,\bold{forcast} means forcast factor return based on historical data,it may take a while,the forcast method come from package \code{\link[forecast]{ets}}.}

\item{covtype}{means type of caculating covariance,\bold{shrink} can see example in \code{\link[nlshrink]{nlshrink_cov}},simple see \code{\link{cov}}.}

\item{nwin}{is rolling windows forward.}

\item{reg_results}{see examples in \code{\link{reg.TSFR}}}

\item{RebDates}{is date set, can be missing.}
}
\value{
a data frame of factors' return .
}
\description{
calculate factor return, factor covariance and residual variance.
}
\examples{
fRtn <- getfRtn(reg_results=reg_results)
fCov <- getfCov(reg_results=reg_results)
Delta <- getDelta(dure=months(1),rolling=FALSE,nwin=24,reg_results)
rtn_cov_delta <- f_rtn_cov_delta(reg_results=reg_results)
}
