% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{factor_na}
\alias{factor_na}
\title{factor_na}
\usage{
factor_na(TSF, method = c("none", "mean", "median"),
  sectorAttr = defaultSectorAttr())
}
\description{
factor_na
}
\examples{
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.NP_YOY(TS)
TSF <- factor_na(TSF, method = "median")
}
\author{
Han.Qian
}
