% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{factor_outlier}
\alias{factor_outlier}
\title{factor_outlier}
\usage{
factor_outlier(TSF, method = c("none", "sd", "mad", "boxplot"), par,
  sectorAttr = defaultSectorAttr())
}
\description{
factor_outlier
}
\examples{
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.NP_YOY(TS)
TSF <- factor_outlier(TSF, method = "mad", par = 1.5)
}
\author{
Han.Qian
}
