% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{factor_result}
\alias{factor_result}
\alias{reg.alpha.select}
\alias{reg.risk.select}
\title{factor select}
\usage{
reg.alpha.select(TSFR)

reg.risk.select(TSFR)
}
\arguments{
\item{TSFR}{a \bold{TSFR} object.}
}
\description{
factor select
}

