% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{factor_select}
\alias{factor_select}
\alias{reg.factor.select}
\title{factor select}
\usage{
reg.factor.select(TSFR, sectorAttr = defaultSectorAttr())
}
\arguments{
\item{TSFR}{a \bold{TSFR} object.}
}
\description{
factor select
}
\examples{
RebDates <- getRebDates(as.Date('2014-01-31'),as.Date('2016-09-30'))
TS <- getTS(RebDates,indexID = 'EI000905')
factorIDs <- c("F000006","F000007","F000008","F000012","F000013","F000014",
"F000015","F000016","F000017","F000018")
tmp <- buildFactorLists_lcfs(factorIDs,factorStd="norm",factorNA = "median")
factorLists <- buildFactorLists(
  buildFactorList(factorFun="gf.NP_YOY",
                  factorPar=list(),
                  factorDir=1),
  buildFactorList(factorFun="gf.ln_mkt_cap",
                  factorPar=list(),
                  factorDir=-1),
  buildFactorList(factorFun="gf.G_MLL_Q",
                  factorPar=list(),
                  factorDir=1),
   buildFactorList(factorFun="gf.GG_NP_Q",
                   factorPar=list(filt = 0),
                   factorDir=1),
   buildFactorList(factorFun="gf.GG_OR_Q",
                   factorPar=list(filt = 0),
                   factorDir=1),                
  factorStd="norm",factorNA = "median")
factorLists <- c(tmp,factorLists)
TSF <- getMultiFactor(TS,FactorLists = factorLists)
TSFR <- getTSR(TSF)
re <- reg.factor.select(TSFR,sectorAttr=NULL)
}

