% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{factorExpDemo}
\alias{factorExpDemo}
\title{factorExpDemo}
\usage{
factorExpDemo(secIn = FALSE)
}
\description{
factor exposure setting demo in optimization function
}
\examples{
fexp <- factorExpDemo()
fexp <- factorExpDemo(secIn = T)
}

