% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{fl2tbl}
\alias{fl2tbl}
\alias{fls2tbl}
\title{fl2tbl}
\usage{
fl2tbl(fl)

fls2tbl(fls)
}
\arguments{
\item{fl}{FactorList}

\item{fls}{FactorLists}
}
\value{
a 2-D table with elements of list class.
}
\description{
turn factorlist or factorlists to 2-D table.
}
\examples{
fl <- buildFactorList(factorFun="gf.pct_chg_per",factorPar=list(N=20,M="test"))
fls <- c(buildFactorLists_lcfs(c("F000001","F000003","F000008","F000009")),list(fl))
re1 <- fl2tbl(fl)
re2 <- fls2tbl(fls)
}
