% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct06_portOptimizing.R
\name{get_exp_rtn}
\alias{get_exp_rtn}
\title{get_exp_rtn}
\usage{
get_exp_rtn(TSF, frtn)
}
\arguments{
\item{frtn}{a vector, or a dataframe}
}
\description{
get_exp_rtn
}
\examples{
# -- frtn as a vector
mtsf <- getMultiFactor_lcfs(ts,c("F000008","F000006"),factorRefine = refinePar_default("scale"))
re <- get_exp_rtn(mtsf,c(0.1,0.3))
# -- frtn as a dataframe
mtsfr <- getTSR(mtsf)
reg_res <- reg.TSFR(mtsfr)
frtn <- getfRtn(reg_results = reg_res)
re2 <- get_exp_rtn(mtsf,frtn)
}
