% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{getbacktestPar}
\alias{getbacktestPar}
\alias{getbacktestPar.IC}
\alias{getbacktestPar.Ngroup}
\alias{getbacktestPar.fee}
\alias{getbacktestPar.longshort}
\alias{getbacktestPar.reg}
\title{getbacktestPar}
\usage{
getbacktestPar(backtestPar, item, item_sub)

getbacktestPar.IC(backtestPar, item_sub)

getbacktestPar.reg(backtestPar, item_sub)

getbacktestPar.Ngroup(backtestPar, item_sub)

getbacktestPar.longshort(backtestPar, item_sub)

getbacktestPar.fee(backtestPar, item_sub)
}
\arguments{
\item{backtestPar}{a \bold{backtestPar} object}

\item{item}{character string, giving the slot name of the backtestPar}

\item{item_sub}{character string, giving the sub-slot name of the specific slot of backtestPar.If missing, then the whole slot will be extracted.}
}
\value{
the slots or sub-slots of the \bold{backtestPar} object.
}
\description{
extract the slot of the \bold{backtestPar} object.
}
\details{
there are two types of extracting methods: \code{getbacktestPar} can extract any slot and any sub-slot, where as \code{getbacktestPar.foo} can extract the foo solt and the sub-slots of foo only.
}
\examples{
backtestPar <- backtestPar.default()
getbacktestPar(backtestPar,"longshort","wgtType")
getbacktestPar(backtestPar,"longshort")
getbacktestPar.longshort(backtestPar,"wgtType")
getbacktestPar.longshort(backtestPar)
}
\author{
Ruifei.Yin
}

