% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{getmodelPar}
\alias{getmodelPar}
\alias{getmodelPar.factor}
\alias{getmodelPar.time}
\alias{getmodelPar.title}
\alias{getmodelPar.univ}
\title{getmodelPar}
\usage{
getmodelPar(modelPar, item, item_sub)

getmodelPar.title(modelPar, item_sub)

getmodelPar.time(modelPar, item_sub)

getmodelPar.univ(modelPar, item_sub)

getmodelPar.factor(modelPar, item_sub)
}
\arguments{
\item{modelPar}{a \bold{modelPar} object}

\item{item}{character string, giving the slot name of the modelPar}

\item{item_sub}{character string, giving the sub-slot name of the specific slot of modelPar.If missing, then the whole slot will be extracted.}
}
\value{
the slots or sub-slots of the \bold{modelPar} object.
}
\description{
extract the slot of the \bold{modelPar} object.
}
\details{
there are two types of extracting methods: \code{getmodelPar} can extract any slot and any sub-slot, where as \code{getmodelPar.foo} can extract the foo solt and the sub-slots of foo only.
}
\examples{
modelPar <- modelPar.default()
getmodelPar(modelPar,"title","modelID")
getmodelPar(modelPar,"title")
getmodelPar.title(modelPar,"modelID")
getmodelPar.title(modelPar)
}
\author{
Ruifei.Yin
}

