% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{getMPs}
\alias{getMPs}
\alias{getMPs_FactorLists}
\alias{getMPs_time.intervals}
\alias{getMPs_time.rebFreqs}
\alias{getMPs_time.shiftbys}
\alias{getMPs_univs}
\title{get modelPar-lists}
\usage{
getMPs_FactorLists(FactorLists, modelPar, nm = laply(FactorLists, function(x)
  x$factorName))

getMPs_time.intervals(intervals, modelPar, nm = paste(intervals[, 1],
  intervals[, 2], sep = "~"))

getMPs_time.rebFreqs(rebFreqs, modelPar, nm = rebFreqs)

getMPs_time.shiftbys(shiftbys, modelPar, nm = as.character(shiftbys))

getMPs_univs(univs, modelPar, nm = univs)
}
\arguments{
\item{FactorLists}{a list of \bold{FactorList}. See more details in \code{\link{buildFactorLists}}.}

\item{modelPar}{a \bold{modelPar} object}

\item{nm}{a character string, specificating the names of the returned list}

\item{intervals}{a dataframe with cols 'begT' and 'endT'. Note that 'begT' must be the 1st col and 'endT' must be the 2nd col.}

\item{rebFreqs}{a charactor vector of \code{rebFreq}. See \code{\link{setmodelPar.time}}}

\item{shiftbys}{a integer vector of \code{shiftby}.}

\item{univs}{a charactor vector of \code{univ}. See \code{\link{setmodelPar.univ}}.}
}
\value{
a \bold{MPs} object, which is a list of \bold{modelPar}
}
\description{
Get ModelPar-lists by the given params. Usually used for mdoel testing with different parameters.
}
\examples{
mp = modelPar.default()

# -- getMPs_FactorLists
factorIDs <- c("F000001","F000002","F000005")
FactorLists <- buildFactorLists_lcfs(factorIDs)
re1 <- getMPs_FactorLists(FactorLists,modelPar=mp)
# -- getMPs_time.intervals
begTs <- seq(as.Date("2010-01-01"),as.Date("2015-01-01"),"year")
endTs <- begTs + months(1)
intervals <- data.frame(begTs,endTs)
re2 <- getMPs_time.intervals(intervals,mp)
# -- getMPs_time.rebFreqs
rebFreqs <- c("day","month","2 month")
re3 <- getMPs_time.rebFreqs(rebFreqs,mp)
# -- getMPs_time.shiftbys
shiftbys <- c(1,3,5,10)
re4 <- getMPs_time.shiftbys(shiftbys,mp)
# -- getMPs_univs
univs <- c("EI000300","EI000905","EI000904")
re5 <- getMPs_univs(shiftbys,mp)
}
\seealso{
\code{\link{Model.TSFs}}, \code{\link{Model.TSFRs}}
}

