% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{getplotPar}
\alias{getplotPar}
\alias{getplotPar.IC}
\alias{getplotPar.MC}
\alias{getplotPar.Ngroup}
\alias{getplotPar.TSFRScatter}
\alias{getplotPar.longshort}
\alias{getplotPar.reg}
\title{getplotPar}
\usage{
getplotPar(plotPar, item, item_sub)

getplotPar.TSFRScatter(plotPar, item_sub)

getplotPar.IC(plotPar, item_sub)

getplotPar.reg(plotPar, item_sub)

getplotPar.Ngroup(plotPar, item_sub)

getplotPar.longshort(plotPar, item_sub)

getplotPar.MC(plotPar, item_sub)
}
\arguments{
\item{plotPar}{a \bold{plotPar} object}

\item{item}{character string, giving the slot name of the plotPar}

\item{item_sub}{character string, giving the sub-slot name of the specific slot of plotPar.If missing, then the whole slot will be extracted.}
}
\value{
the slots or sub-slots of the \bold{plotPar} object.
}
\description{
extract the slot of the \bold{plotPar} object.
}
\details{
there are two types of extracting methods: \code{getplotPar} can extract any slot and any sub-slot, where as \code{getplotPar.foo} can extract the foo solt and the sub-slots of foo only.
}
\examples{
plotPar <- plotPar.default()
getplotPar(plotPar,"Ngroup","Nbin")
getplotPar(plotPar,"Ngroup")
getplotPar.Ngroup(plotPar,"Nbin")
getplotPar.Ngroup(plotPar)
}
\author{
Ruifei.Yin
}

