% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct06_portOptimizing.R
\name{getPort_opt}
\alias{getPort_opt}
\title{getPort_opt}
\usage{
getPort_opt(TSF, bmk = NULL, exp_rtn = "factorscore", risk_list = NULL,
  lamda = NULL, constr = constr_default(), init_port, min_wgt = 1e-04,
  optsolver = c("CVXR", "RMOSEK"))
}
\arguments{
\item{TSF}{a \bold{TSF} object,may contains multiple factors.}

\item{bmk}{benchmark indexID code.}

\item{exp_rtn}{a string, indicating the expected rtn column name. Or a numeric 0, which means return part doesn't involved in the optimization.}

\item{lamda}{numeric scalar, risk_aversion}

\item{constr}{constrain lists,see \code{\link{opt_constrain}}.}

\item{init_port}{if turnover constrain involved ,param \code{init_port} is required.}

\item{min_wgt}{stock's minimum weight in optimized portfolio.}
}
\value{
a optimized port
}
\description{
getPort_opt
}
\examples{
TS <- getTS(RebDates = as.Date("2017-03-31"),indexID = "EI000300")
TSF <- getTSF(TS,FactorList = buildFactorList_lcfs("F000006",factorRefine = refinePar_default("scale")))
# constrain setting
constr <- constr_default(box_each = c(0,0.02))
constr <- addConstr_box(constr,each = c(0,0.02))
constr <- addConstr_fctExp_sector(constr,each = c(-0.05,0.05))
conslist <- buildFactorLists_lcfs("F000002",factorRefine = refinePar_default("scale",NULL))
# with bmk
constr <- addConstr_fctExp_style(constr,conslist,-0.1,0.1)
port_opt <- getPort_opt(TSF,bmk = "EI399330",constr = constr,exp_rtn = 'factorscore')
# without bmk
constr2 <- setConstr_fctExp_sector(constr,each = c(0,0.1),relative = 0)
port_opt2 <- getPort_opt(TSF,bmk=NULL,constr = constr2,exp_rtn = 'factorscore')
# long-short port
constr3 <- setConstr_position(constr_default(),position = c(0,0),relative = 0)
constr3 <- setConstr_box(constr3,each=c(-0.05,0.05),relative=0)
port_opt3 <- getPort_opt(TSF,bmk="EI000300",constr = constr3,exp_rtn = 'factorscore')
}
