% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct06_portOptimizing.R
\name{getPort_opt_old}
\alias{getPort_opt_old}
\title{getPort_opt_old}
\usage{
getPort_opt_old(TSF, alphaf, fRtn = data.frame(fname = alphaf, frtn =
  rep(1/length(alphaf), length(alphaf)), stringsAsFactors = FALSE), fCov,
  exp_rtn = "exp_rtn", bmk = NULL, constr = constr_default(),
  addEvent = FALSE, target = c("return", "balance"), optWay = c("ipop",
  "solve.QP", "Matlab"))
}
\description{
getPort_opt_old
}
