% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{getRAData}
\alias{getRAData}
\title{getRAData}
\usage{
getRAData(port, factorLists, bmk, sectorAttr = defaultSectorAttr())
}
\description{
getRAData
}
\examples{
alphaLists <- buildFactorLists_lcfs(c("F000012","F000008"),factorRefine=refinePar_default("scale"))
alphaLists <- c(tmp,alphaLists)
riskLists <- buildFactorLists_lcfs(c("F000002","F000006"),factorRefine=refinePar_default("none"))
RA_tables <- getRAData(port,c(alphaLists,riskLists))
RA_tables <- getRAData(port,c(alphaLists,riskLists),bmk='EI000905')
}
