% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{getRebDates}
\alias{getRebDates}
\title{getRebDates}
\usage{
getRebDates(begT, endT, rebFreq = "month", shiftby = 0, dates = NULL)
}
\arguments{
\item{begT}{the begin date of class Date}

\item{endT}{the end date of class Date}

\item{rebFreq}{the rebalancing freq. an interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s". (See \code{\link{cut.Date}} for detail.) Or a character string of "allspan". "allspan" means no rebalance during the whole period between begT and endT, which could be realized also by a very large interval specification, eg. "100 year".}

\item{shiftby}{a integer,how many days the rebalancing date be shifted afterward}

\item{dates}{a vector of class Date.}
}
\value{
a \bold{RebDates} object giving the rebalancing dates list,a vector with class of \code{Date}. If dates is not null, then return the dates themselves.
}
\description{
get the rebalance dates(a \bold{RebDates} object) between begT and endT, and then shift it afterward if necessary.
}
\examples{
RebDates <- getRebDates(as.Date('2010-01-01'),as.Date('2012-06-30'),'3 month', 10)
getRebDates(as.Date('2010-01-01'),as.Date('2012-06-30'),'allspan')
getRebDates(as.Date('2010-01-01'),as.Date('2012-06-30'),'100 year')
getRebDates(dates=as.Date(c("2003-04-09","2003-07-06","2003-09-12")))
}
\seealso{
Other frame building functions: \code{\link{getTSF}},
  \code{\link{getTSR}}, \code{\link{getTS}}
}
\author{
Ruifei.Yin
}
