% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{getResidual}
\alias{getResidual}
\title{getResidual}
\usage{
getResidual(TS, dure, reg_results)
}
\arguments{
\item{TS}{a \bold{TS} object.}

\item{dure}{a period object from package \code{lubridate}. (ie. \code{months(1),weeks(2)}. See example in \code{\link{trday.offset}}.) If null, then get periodrtn between \code{date} and the next \code{date}, else get periodrtn of '\code{dure}' starting from \code{date}.}
}
\description{
get stocks' residual
}
\examples{
RebDates <- getRebDates(as.Date('2012-01-31'),as.Date('2016-08-31'))
TS <- getTS(RebDates)
dure <- lubridate::days(1)
res <- getResidual(TS,dure)
dure <- months(1)
res <- getResidual(TS,dure)
}

