% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct02_portBuildingFuncs.R
\name{getrtn.bmk}
\alias{getrtn.bmk}
\title{getrtn.bmk}
\usage{
getrtn.bmk(rtn, bmk = "EI000300")
}
\arguments{
\item{rtn}{an xts,timeSeries or zoo object of portfolio returns,which must be 1 column}

\item{bmk}{a character string, giving the stockID of the benchmark index,eg. "EI000300".}
}
\value{
a xts object, giving the rtn series of the benchmark
}
\description{
get the benchmark return series corresponding to the specific rtn series.
}
\examples{
rtn <- zoo(rnorm(1000,0.001,0.02),as.Date("2010-01-01")+1:1000)
re <- getrtn.bmk(rtn,"EI000300")
}
\author{
Ruifei.Yin
}
\seealso{
Other LSH-frame building functions: \code{\link{addrtn.hedge}},
  \code{\link{getrtn.LBH}}, \code{\link{getrtn.LSH}}
}

