% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct02_portBuildingFuncs.R
\name{getrtn.LBH}
\alias{getrtn.LBH}
\title{getrtn.LBH}
\usage{
getrtn.LBH(PB.L, bmk = "EI000300", hedge.rebFreq = "month",
  hedge.posi = 1, fee.long = 0, fee.short = 0, backtestPar)
}
\arguments{
\item{PB.L}{a \bold{PB} object of the long portfolio}

\item{bmk}{a character string,giving the stockID of the benchmark index,eg. "EI000300".}

\item{hedge.rebFreq}{giving the rebalance freq when computing the hedged rtn.An interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s".See \code{\link{cut.Date}} for detail.}

\item{hedge.posi}{a numeric, giving the position of the hedging assets}

\item{fee.long}{a numeric, giving the fee of long asset}

\item{fee.short}{a numeric, giving the fee of short asset}

\item{backtestPar}{a \bold{backtestPar} object}
}
\value{
a \bold{rtn.LBH} object of class xts,giving the return series of long,bmk and hedge.
}
\description{
get the \bold{rtn.LBH} object from the \bold{TSF} object.
}
\details{
If param backtestPar is not missing,then the related params will be extracted from backtestPar.It is usefull when the backtestPar has been initialized.
}
\examples{
re <- getrtn.LBH(PB.L,"EI000300")
}
\author{
Ruifei.Yin
}
\seealso{
Other LSH-frame building functions: \code{\link{addrtn.hedge}},
  \code{\link{getrtn.LSH}}, \code{\link{getrtn.bmk}}
}

