% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs-³åÍ»-ÀÏÈý.R, R/fct01_frameBuildingFuncs.R
\name{getTS}
\alias{getTS}
\title{getTS}
\usage{
getTS(RebDates, indexID = "EI000300", stocks = NULL)

getTS(RebDates, indexID = "EI000300", stocks = NULL)
}
\arguments{
\item{RebDates}{a \bold{RebDates} object. a vector,with class of Date,usually the rebalancing dates list}

\item{indexID}{a character string. The ID of the index, sector or plate. Could be a single-ID-code(eg. "EI000300","ES09440000",...) or a more complicated express containing some set operations and ID-codes(eg. "setdiff(union(EI000300,EI399006),ES09440000)"). See detail in \code{\link{getComps}}.}

\item{stocks}{a vector of stockIDs}

\item{RebDates}{a \bold{RebDates} object. a vector,with class of Date,usually the rebalancing dates list}

\item{indexID}{a character string. The ID of the index, sector or plate. Could be a single-ID-code(eg. "EI000300","ES09440000",...) or a more complicated express containing some set operations and ID-codes(eg. "setdiff(union(EI000300,EI399006),ES09440000)"). See detail in \code{\link{getComps}}.}

\item{stocks}{a vector of stockIDs}
}
\value{
a \bold{TS} object. a dataframe,with cols:
  \itemize{
  \item date: the rebalance dates, with \code{Date} class
  \item stockID: the stockID 
  }.IF stocks is not null then return \code{expand.grid} of stocks and RebDates, else return components of index on RebDates.

a \bold{TS} object. a dataframe,with cols:
  \itemize{
  \item date: the rebalance dates, with \code{Date} class
  \item stockID: the stockID 
  }.IF stocks is not null then return \code{expand.grid} of stocks and RebDates, else return components of index on RebDates.
}
\description{
get the 'time*stock frame'(a \bold{TS} object) given the rebalancing dates and the universe. The universe could be a index, a sector or a plate.

get the 'time*stock frame'(a \bold{TS} object) given the rebalancing dates and the universe. The universe could be a index, a sector or a plate.
}
\examples{
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
TS2 <- getTS(RebDates,'ES09440000')  # financial servive sector
TS3 <- getTS(RebDates,'setdiff(EI000300,ES09440000)')  # CSI300 ex. financial servive sector
TS4 <- getTS(RebDates,stocks=c("EQ000002","EQ000023","EQ600000"))
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
TS2 <- getTS(RebDates,'ES09440000')  # financial servive sector
TS3 <- getTS(RebDates,'setdiff(EI000300,ES09440000)')  # CSI300 ex. financial servive sector
TS4 <- getTS(RebDates,stocks=c("EQ000002","EQ000023","EQ600000"))
}
\author{
Ruifei.Yin

Ruifei.Yin
}
\seealso{
\code{\link{getComps}}

\code{\link{getComps}}

Other frame building functions: \code{\link{getRebDates}},
  \code{\link{getTSF}}, \code{\link{getTSR}}

Other frame building functions: \code{\link{getRebDates}},
  \code{\link{getTSF}}, \code{\link{getTSR}}
}

