% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{getTSR}
\alias{getTSR}
\title{getTSR}
\usage{
getTSR(TS, dure = NULL, date_end_pad)
}
\arguments{
\item{TS}{a \bold{TS} or \bold{TSF} object (See detail in \code{\link{getTS}} and \code{\link{getTSF}})}

\item{dure}{a period object from package \code{lubridate}. (ie. \code{months(1),weeks(2)}. See example in \code{\link{trday.offset}}.) If null, then get periodrtn between \code{date} and the next \code{date}, else get periodrtn of '\code{dure}' starting from \code{date}.}

\item{date_end_pad}{a Date value, padding the NAs of the last date_end. if missing, calculated automatically.}
}
\value{
given a \bold{TS} object,return a \bold{TSR} object,a dataframe containing cols:
  \itemize{ 
  \item stockID: the stockID contained in the index or BK
  \item date: the rebalance dates, with \code{Date} class 
  \item date_end: the next rebalance dates, with \code{Date} class
  \item periodrtn:the period rtn between the rebalance date and the date_end
  }
  given a \bold{TSF} object,return a \bold{TSFR} object,a dataframe containing cols:
  \itemize{ 
  \item stockID: the stockID contained in the index or BK
  \item date: the rebalance dates, with \code{Date} class
  \item factorscore:the factor score of the stocks on the  rebalancing date  
  \item date_end: the next rebalance dates, with \code{Date} class
  \item periodrtn:the period rtn between the rebalance date and the date_end
  }
}
\description{
Get the \bold{TSR}('time-stock-rtn') or \bold{TSFR}('time-stock-factor-rtn') object by adding the period rtn between the rebalance date and the date_end(\strong{(close of "date_end")/(prevclose of the day after "date")-1)}) to the \bold{TS} or \bold{TSF} object.
}
\examples{
## -- get a TSR object
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300') 
TSR <- getTSR(TS) # rebalance properly joined
require(lubridate)
TSR2 <- getTSR(TS,weeks(2)); TSR3 <- getTSR(TS,months(2))  # rebalance not properly joined
## -- get a TSFR object
factorFun <- "gf.demo"
factorPar <- list(0,1)
TSF <- getTSF(TS,factorFun,factorPar)
TSFR <-  getTSR(TSF)
}
\author{
Ruifei.Yin
}
\seealso{
Other frame building functions: \code{\link{getRebDates}},
  \code{\link{getTSF}}, \code{\link{getTS}}
}

