% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs-³åÍ»-ÀÏÈý.R, R/fct01_frameBuildingFuncs.R
\name{getTSR_decay}
\alias{getTSR_decay}
\title{getTSR_decay}
\usage{
getTSR_decay(TS, prd_lists = list(w1 = lubridate::weeks(1), w2 =
  lubridate::weeks(2), m1 = months(1), m2 = months(2), m3 = months(3), m6 =
  months(6)))

getTSR_decay(TS, prd_lists = list(w1 = lubridate::weeks(1), w2 =
  lubridate::weeks(2), m1 = months(1), m2 = months(2), m3 = months(3), m6 =
  months(6)))
}
\value{
a \bold{TSR} including the decayed period rtns

a \bold{TSR} including the decayed period rtns
}
\description{
getTSR_decay

getTSR_decay
}

