% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{gf_demo}
\alias{gf_demo}
\title{A demo function of factor-getting}
\usage{
gf_demo(TS, rho = 0, dure = NULL, date_end_pad)
}
\arguments{
\item{rho}{a numeric. the desired correlation between Periodrtn and factorscore. See \code{\link{[QUtilily]getBiCop}}}
}
\value{
a TSF object
}
\description{
You can get different result by adjusting the param \code{rho}.
}
\author{
Ruifei.Yin
}
