% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{ggplot.corplot}
\alias{ggplot.corplot}
\title{ggplot.corplot}
\usage{
\method{ggplot}{corplot}(corr)
}
\arguments{
\item{corr}{is correlation matrix or a list of correlation matrix ,see \code{\link[stats]{cor}}.}
}
\description{
ggplot.corplot
}
