% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{lcdb_regtables}
\alias{lcdb.build.RegTables}
\alias{lcdb.update.RegTables}
\alias{lcdb_regtables}
\title{lcdb_regtables}
\usage{
lcdb.build.RegTables(FactorLists, begT = 20050104, endT)

lcdb.update.RegTables(FactorLists)
}
\description{
build and update local database's regression result tables
}
\examples{
FactorLists <- buildFactorLists_lcfs(factorIDs = c("F000002","F000006","F000008","F000017"),factorStd = "norm",factorNA = "median")
lcdb.build.RegTables(FactorLists,20150101,20161231)
lcdb.update.RegTables(FactorLists)
}

