% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{MC.table.fCorr}
\alias{MC.table.fCorr}
\title{MC.table.fCorr}
\usage{
MC.table.fCorr(TSF, Nbin)
}
\description{
MC.table.fCorr
}
\examples{
factorIDs <- c("F000006","F000008","F000012")
FactorLists <- buildFactorLists_lcfs(factorIDs,factorStd="norm",factorNA = "mean")
TSF <- getMultiFactor(TS,FactorLists = FactorLists)
MC.table.fCorr(TSF)
MC.table.fCorr(TSF,Nbin='year')
}

