% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{MC.wgt.CAPM}
\alias{MC.wgt.CAPM}
\title{MC.wgt.CAPM}
\usage{
MC.wgt.CAPM(TSFRs, stat = c("pearson", "spearman"), backtestPar)
}
\value{
a vector
}
\description{
compute the wgt vector of multi-factors by CAPM model.
}
\examples{
mp = modelPar.default()
factorIDs <- c("F000001","F000002","F000005")
FactorLists <- buildFactorLists_lcfs(factorIDs)
mps <- getMPs_FactorLists(FactorLists,modelPar=mp)
TSR <- Model.TSR(mp)
TSFRs <- Model.TSFs_byTS(MPs=mps,TS=TSR)
MC.wgt.CAMP(TSFRs) 
}

