% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct04_modelMethods.R
\name{ModelMethods}
\alias{ModelMethods}
\alias{Model.RebDates}
\alias{Model.TS}
\alias{Model.TSR}
\alias{Model.TSF}
\alias{Model.TSF_byTS}
\alias{Model.TSFR}
\alias{Model.TSFs}
\alias{Model.TSFRs}
\alias{Model.TSFs_byTS}
\title{Model methods}
\usage{
Model.RebDates(modelPar)

Model.TS(modelPar)

Model.TSR(modelPar)

Model.TSF(modelPar)

Model.TSF_byTS(modelPar, TS)

Model.TSFR(modelPar)

Model.TSFs(MPs, nm = names(MPs))

Model.TSFRs(MPs, nm = names(MPs))

Model.TSFs_byTS(MPs, TS, nm = names(MPs))
}
\arguments{
\item{modelPar}{a \bold{modelPar} object}

\item{TS}{a \bold{TS} or \bold{TSR} object}

\item{MPs}{}

\item{nm}{a character string, specificating the names of the returned list}
}
\value{
\code{Model.RebDates} return a \bold{RebDates} object(See detail in \code{\link{getRebDates}}).

\code{Model.TS} return a \bold{TS} object(See detail in \code{\link{getTS}}).

\code{Model.TSR} return a \bold{TSR} object(See detail in \code{\link{getTSR}})

\code{Model.TSF} get a \bold{TSF} object by modelPar directly.(See detail in \code{\link{getTSF}})

\code{Model.TSF_byTS} get the \bold{TSF}(or \code{TSFR}) object by modelPar and \code{TS}(or \code{TSR}).

\code{Model.TSFR} return a \bold{TSFR} object(See detail in \code{\link{getTSR}} and \code{\link{getTSF}})

\code{Model.TSFs} return a named list of \bold{TSF} object, whose elements is coresponding to the elements of \bold{MPs}

\code{Model.TSFRs} return a named list of \bold{TSFR} object, whose elements is coresponding to the elements of \bold{MPs}

\code{Model.TSFs_byTS} return a named list of \bold{TSF}(or \code{TSFR}) object, whose elements is coresponding to the elements of \bold{MPs}
}
\description{
Get the objects of \code{RebDates}, \code{TS}, \code{TSF}, \code{TSR}, \code{TSFR}, \code{TSFRM} through a \code{modelPar} object.
}
\note{
Function \code{Model.TSF_byTS} is offen used when getting \code{TS} is time costly, or when test different factor related parametres frequently. Note that if parametre \code{TS} is a \code{TSR} object, then \code{Model.TSF_byTS} return a \code{TSFR} object, else return a \code{TSF} object.

Function \code{Model.TSFs_byTS} is offen used when getting \code{TS} is time costly, or when test different factor related parametres frequently. Note that if parametre \code{TS} is a \code{TSR} object, then \code{Model.TSFs_byTS} return a list of \code{TSFR} object, else return a list of \code{TSF} object.
}
\examples{
modelPar <- modelPar.default()

# -- get the RebDates object
RebDates <- Model.RebDates(modelPar)
# -- get the TS object
TS <- Model.TS(modelPar)
# -- get the TSR object
TSR <- Model.TSR(modelPar)
# -- get the TSF object
TSF <- Model.TSF(modelPar)
# -- get the TSF object by TS
TS <- Model.TS(modelPar)
TSF <- Model.TSF_byTS(modelPar,TS)
TSF2 <- Model.TSF_byTS(setmodelPar.factor(modelPar,factorOutlier = 10),TS)
# -- get the TSFR object by TSR
TSR <- Model.TSR(modelPar)
TSFR <- Model.TSF_byTS(modelPar,TSR)
TSFR2 <- Model.TSF_byTS(setmodelPar.factor(modelPar,factorOutlier = 10),TSR)
# -- get the TSFR object
TSFR <- Model.TSFR(modelPar)
# -- get the TSFs list
factorIDs <- c("F000001","F000002","F000005")
FactorLists <- buildFactorLists_lcfs(factorIDs)
MPs <- getMPs_FactorLists(FactorLists, modelPar.default())
TSFs <- Model.TSFs(MPs)
# -- get the TSFRs list
TSFRs <- Model.TSFRs(MPs)
# -- get the list by TS and TSR
TSFs <- Model.TSFs_byTS(MPs, TS)
TSFRs <- Model.TSFs_byTS(MPs, TSR)
}
\author{
Ruifei.Yin
}
