% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{modelPar.default}
\alias{modelPar.default}
\title{modelPar.default}
\usage{
modelPar.default()
}
\value{
a \bold{modelPar} object with default parametres
}
\description{
get a \bold{modelPar} object with default parametres
}
\author{
Ruifei.Yin
}
\seealso{
Other modelPar setting functions: \code{\link{modelPar.factor}},
  \code{\link{modelPar.time}},
  \code{\link{modelPar.title}}, \code{\link{modelPar.univ}}
}

