% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{modelPar.factor}
\alias{modelPar.factor}
\alias{setmodelPar.factor}
\alias{setmodelPar.factor_combi}
\title{modelPar.factor}
\usage{
modelPar.factor(modelPar = modelPar.default(), factorFun = "gf_demo",
  factorPar = list(), factorDir = 1,
  factorRefine = refinePar_default("none"),
  factorName = default.factorName(factorFun, factorPar, factorDir),
  factorID = "", factorType = "", factorDesc = "", FactorList)

setmodelPar.factor(modelPar, factorFun, factorPar, factorDir, factorRefine,
  factorName, factorID, factorType, factorDesc, FactorList)

setmodelPar.factor_combi(modelPar, factorLists, wgts, factorDir, factorRefine,
  factorName = "combi_factor", factorID, factorType, factorDesc)
}
\arguments{
\item{modelPar}{a \bold{modelPar} object}

\item{factorFun}{a non-empty character string naming the function to get the factor scores.(See \code{\link{getTSF}})}

\item{factorPar}{a list, or a character string, containing the parameters of the \code{factorFun}.(See \code{\link{getTSF}})}

\item{factorDir}{a integer,should be 1 or -1 (1 for the positive factor,-1 for the negative one).(See \code{\link{getTSF}})}

\item{factorRefine}{a list. See detail in \code{\link{refinePar_default}}}

\item{factorName}{a character string. IF missing, then take a default name by function \code{default.factorName}.}

\item{FactorList}{a list containing all the factor related parametres in it. See more detail in \code{\link{buildFactorList}}.}

\item{factorLists}{See \code{\link{buildFactorLists}}.}
}
\value{
a \bold{modelPar} object.
}
\description{
set the factor related parametres of the model
}
\details{
The different between \code{modelPar.factor} and \code{setmodelPar.factor} is that the all pars of the former have default values, and the latter have not. The former is ussually used to to build the modelPar integrally, and the latter is ussually used to set the parametre individually.

if \code{FactorList} is missing, then the factor related parametres is set by the auguments separately, else, all the parametres is set by the \code{FactorList} once and for all.
}
\note{
if the parammeter \code{factorFun} is \code{"gf_lcfs"}, then the factorDir, factorName, factorID, factorType and factorDesc will be mapped automatically by \code{\link{CT_FactorLists}}.

\code{setmodelPar.factor_combi} set \code{modelPar} by \code{factorFun="getMultiFactor"} and  \code{factorPar=list(factorLists,wgts)}.  It is used usually in testing the multi-factor-model. See \code{\link{getMultiFactor}}.
}
\examples{
modelPar <- modelPar.default()
modelPar <- modelPar.factor(modelPar,
                           factorFun="gf.pct_chg_per",
                           factorPar=list(N=25),
                           factorDir=-1,
                           factorRefine= setrefinePar(refinePar_default("scale"))
# -- setmodelPar.factor_combi
mp <- modelPar.default()
factorIDs <- c("F000008","F000001","F000006")
factorLists <- buildFactorLists_lcfs(factorIDs)
wgts <- c(0.5, 0.3, 0.2)
mp_m <- setmodelPar.factor_combi(mp, factorLists, wgts)
# -- setmodelPar.factor_combi by \\code{buildFactorlist_combi}
factorlist_combi <- buildFactorList_combi(factorLists,wgts)
mp_m2 <- setmodelPar.factor(mp,FactorList = factorlist_combi)
}
\seealso{
\code{\link{getTSF}}

Other modelPar setting functions: \code{\link{modelPar.default}},
  \code{\link{modelPar.time}},
  \code{\link{modelPar.title}}, \code{\link{modelPar.univ}}
}
\author{
Ruifei.Yin
}
