% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{modelPar.title}
\alias{modelPar.title}
\alias{setmodelPar.title}
\title{modelPar.title}
\usage{
modelPar.title(modelPar = modelPar.default(), modelID = "",
  modelName = "", modelDesc = "")

setmodelPar.title(modelPar, modelID, modelName, modelDesc)
}
\arguments{
\item{modelPar}{a \bold{modelPar} object}

\item{modelID}{}
}
\value{
a \bold{modelPar} object
}
\description{
set the title related parametres of the model
}
\details{
The different between \code{modelPar.title} and \code{setmodelPar.title} is that the all pars of the former have default values, and the latter have not. The former is ussually used to to build the modelPar integrally, and the latter is ussually used to set the parametre individually.
}
\examples{
modelPar <- modelPar.default()
modelPar.title(modelPar,modelID="M_000001",modelName="HS300-eqwgt-momentumn")
}
\seealso{
Other modelPar setting functions: \code{\link{modelPar.default}},
  \code{\link{modelPar.factor}},
  \code{\link{modelPar.time}}, \code{\link{modelPar.univ}}
}
\author{
Ruifei.Yin
}
