% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{modelPar.univ}
\alias{modelPar.univ}
\alias{setmodelPar.univ}
\title{modelPar.univ}
\usage{
modelPar.univ(modelPar = modelPar.default(), indexID = "EI000300",
  stocks = NULL, rm = NULL)

setmodelPar.univ(modelPar, indexID, stocks, rm)
}
\arguments{
\item{modelPar}{a \bold{modelPar} object}

\item{indexID}{the ID of the index("EIxxxx"), sector("ESxxxx"), or plate("EPxxxx"). The function will recognize the universe type through the first two character of \code{indexID}.}

\item{stocks}{}

\item{rm}{}
}
\value{
a \bold{modelPar} object
}
\description{
set the universe related parametres of the model
}
\details{
The different between \code{modelPar.univ} and \code{setmodelPar.univ} is that the all pars of the former have default values, and the latter have not. The former is ussually used to to build the modelPar integrally, and the latter is ussually used to set the parametre individually.
}
\seealso{
\code{\link{getTS}}

Other modelPar setting functions: \code{\link{modelPar.default}},
  \code{\link{modelPar.factor}},
  \code{\link{modelPar.time}}, \code{\link{modelPar.title}}
}
\author{
Ruifei.Yin
}
