% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct06_portOptimizing.R
\name{opt_object}
\alias{opt_object}
\alias{object_default}
\title{opt_object}
\usage{
object_default()
}
\value{
a list contain the constrains
}
\description{
build, set, add, clear objects
}
\examples{
obj <- object_default()
}
