% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{OptWgt}
\alias{OptWgt}
\title{OptWgt}
\usage{
OptWgt(TSF, alphaf, fRtn, fCov, target = c("return", "balance"), bmk,
  sectorAttr = defaultSectorAttr(), factorExp = buildFactorExp(),
  wgtSet = buildWgtSet(), boxConstr, addEvent = FALSE, optWay = c("ipop",
  "solve.QP", "Matlab"))
}
\arguments{
\item{TSF}{is multiple factors' \bold{TSF} object}

\item{alphaf}{is alpha factors' name, can be missing.}

\item{fRtn}{see \code{\link{getfRtn}}}

\item{fCov}{is the covariance matrix.}

\item{bmk}{is the benckmark for optimization,can be missing.}
}
\value{
a \bold{port} object.
}
\description{
optimize portfolio weight.
}
\examples{


}
\author{
Andrew Dow
}
