% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{OptWgt_settingFuncs}
\alias{OptWgt_settingFuncs}
\alias{buildFactorExp}
\alias{buildWgtSet}
\alias{buildBoxConstr}
\title{OptWgt_settingFuncs}
\usage{
buildFactorExp(sectorall = c(-0.05, 0.05), ...)

buildWgtSet(wgtall = c(0, 0.01), ...)

buildBoxConstr(...)
}
\description{
\bold{buildFactorExp} build factor exposure setting in optimization function.
\bold{buildWgtSet} build weight setting in optimization function.
\bold{buildBoxConstr} build box constraint in optimization function.
}
\examples{
fexp <- buildFactorExp()
fexp <- buildFactorExp(disposition_=c(-0.01,1),beta_ = c(-0.01, 1))
fexp <- buildFactorExp(sectorall=c(-0.05,0.05),disposition_=c(-0.01,1),ES33480000=c(0.15,0.25))
---------------------------------------------------------------------
wgtSet <- buildWgtSet()
wgtSet <- buildWgtSet(wgtall=c(0,0.02),EQ601318=c(0.05,0.15))
wgtSet <- buildWgtSet(EQ601318=c(0.05,0.15),ES33370000=c(0,0.5))
---------------------------------------------------------------------
boxConstr <- buildBoxConstr(EI000905=c(0.75,0.85))
boxConstr <- buildBoxConstr(EI000905=c(0.75,0.85),EI000300=c(0.15,0.35))
}
