% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{PA_RA_Analysis}
\alias{PA_RA_Analysis}
\alias{getPAData}
\alias{chart.PA.exposure}
\alias{chart.PA.attr}
\alias{chart.RA.attr}
\title{PA_RA_Analysis}
\usage{
getPAData(port, FactorLists, bmk, sectorAttr = defaultSectorAttr())

chart.PA.exposure(PA_tables, plotInd = FALSE)

chart.PA.attr(PA_tables, plotInd = FALSE, attributeAnn = TRUE)

chart.RA.attr(PA_tables)
}
\description{
performance attribution and risk attribution analysis.

getPAData

chart.PA.exposure

chart.PA.attr

chart.RA.attr
}
\examples{
FactorLists <- buildFactorLists(
  buildFactorList(factorFun="gf.SIZE"),
  buildFactorList(factorFun="gf.GROWTH"),
  buildFactorList(factorFun="gf.TRADING"),
  buildFactorList(factorFun="gf.FORECAST"),
  buildFactorList(factorFun="gf.EARNINGYIELD"),
  buildFactorList(factorFun="gf.VALUE"),
  buildFactorList(factorFun="gf.QUALITY"))
PA_tables <- getPAData(port,FactorLists)
PA_tables <- getPAData(port,FactorLists,bmk='EI000905')
chart.PA.exposure(PA_tables)
chart.PA.exposure(PA_tables,plotInd=TRUE)
chart.PA.attr(PA_tables)
chart.PA.attr(PA_tables,plotInd=TRUE)
chart.RA.attr(PA_tables)
}
