% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{plotPar.default}
\alias{plotPar.default}
\title{plotPar.default}
\usage{
plotPar.default()
}
\value{
a \bold{plotPar} object with default parametres
}
\description{
set the plotting with default parametres
}
\author{
Ruifei.Yin
}
\seealso{
Other plotPar setting functions: \code{\link{plotPar.IC}},
  \code{\link{plotPar.MC}}, \code{\link{plotPar.Ngroup}},
  \code{\link{plotPar.TSFRScatter}},
  \code{\link{plotPar.longshort}}
}

