% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{plotPar.IC}
\alias{plotPar.IC}
\alias{setplotPar.IC}
\title{plotPar.IC}
\usage{
plotPar.IC(plotPar = plotPar.default(), Nbin = "day", stat = c("pearson",
  "spearman"))

setplotPar.IC(plotPar, Nbin, stat)
}
\arguments{
\item{plotPar}{a \bold{plotPar} object}

\item{Nbin}{the number of the groups the timespan is cut to, when plotting the IC series.It could also be character of interval specification,See \code{\link{cut.Date}} for detail.}

\item{stat}{a character string,indicating the methods to compute IC,could be "pearson" or "spearman".}
}
\value{
a \bold{plotPar} object
}
\description{
set the IC related parametres of the ploting
}
\seealso{
\code{\link{chart.IC}}

Other plotPar setting functions: \code{\link{plotPar.MC}},
  \code{\link{plotPar.Ngroup}},
  \code{\link{plotPar.TSFRScatter}},
  \code{\link{plotPar.default}},
  \code{\link{plotPar.longshort}}
}
\author{
Ruifei.Yin
}
