% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{plotPar.longshort}
\alias{plotPar.longshort}
\alias{setplotPar.longshort}
\title{plotPar.longshort}
\usage{
plotPar.longshort(plotPar = plotPar.default(), bar.freq = "month",
  roll.width = 250, roll.by = 30)

setplotPar.longshort(plotPar, bar.freq, roll.width, roll.by)
}
\arguments{
\item{plotPar}{a \bold{plotPar} object}

\item{bar.freq}{the freq of the per-period performance bar chart}

\item{roll.width}{the width argument for rolling performance chart}

\item{roll.by}{the by argument for rolling performance chart}
}
\value{
a \bold{plotPar} object
}
\description{
set the long-short related parametres of the ploting
}
\seealso{
\code{\link{plot.longshort}}

Other plotPar setting functions: \code{\link{plotPar.IC}},
  \code{\link{plotPar.MC}}, \code{\link{plotPar.Ngroup}},
  \code{\link{plotPar.TSFRScatter}},
  \code{\link{plotPar.default}}
}
\author{
Ruifei.Yin
}
