% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{plotPar.MC}
\alias{plotPar.MC}
\alias{setplotPar.MC}
\title{plotPar.MC}
\usage{
plotPar.MC(plotPar = plotPar.default(), ncol.IC = 3, ncol.IC.decay = 3,
  ncol.Ngroup = 3)

setplotPar.MC(plotPar, ncol.IC, ncol.IC.decay, ncol.Ngroup)
}
\arguments{
\item{plotPar}{a \bold{plotPar} object}

\item{ncol.IC}{}
}
\value{
a \bold{plotPar} object
}
\description{
set the multi-comparison related parametres of the ploting
}
\seealso{
Other plotPar setting functions: \code{\link{plotPar.IC}},
  \code{\link{plotPar.Ngroup}},
  \code{\link{plotPar.TSFRScatter}},
  \code{\link{plotPar.default}},
  \code{\link{plotPar.longshort}}
}
\author{
Ruifei.Yin
}
