% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{plotPar.Ngroup}
\alias{plotPar.Ngroup}
\alias{setplotPar.Ngroup}
\title{plotPar.Ngroup}
\usage{
plotPar.Ngroup(plotPar = plotPar.default(), Nbin = "day", N = 5,
  stat = c("mean", "median"), turnoverType = c("num", "wgt"))

setplotPar.Ngroup(plotPar, Nbin, N, stat, turnoverType)
}
\arguments{
\item{plotPar}{a \bold{plotPar} object}

\item{Nbin}{the number of the groups the timespan is cut to, when plotting the "date.grp".It could also be character of interval specification,See \code{\link{cut.Date}} for detail.}

\item{N}{the number of the groups the universe is cut to}

\item{stat}{a character string,indicating the statistic of the return center of each group,could be "mean" or "median".}

\item{turnoverType}{a character string,indicating the method to calculate the turnover,could be "num" or "wgt".}
}
\value{
a \bold{plotPar} object
}
\description{
set the Ngroup related parametres of the ploting
}
\seealso{
\code{\link{chart.Ngroup.overall}},\code{\link{chart.Ngroup.seri}}

Other plotPar setting functions: \code{\link{plotPar.IC}},
  \code{\link{plotPar.MC}},
  \code{\link{plotPar.TSFRScatter}},
  \code{\link{plotPar.default}},
  \code{\link{plotPar.longshort}}
}
\author{
Ruifei.Yin
}
