% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct00_modelsettingFuncs.R
\name{plotPar.TSFRScatter}
\alias{plotPar.TSFRScatter}
\title{plotPar.TSFRScatter}
\usage{
plotPar.TSFRScatter(plotPar = plotPar.default(), Nbin = "day")
}
\arguments{
\item{plotPar}{a \bold{plotPar} object}

\item{Nbin}{the number of the groups the timespan is cut to when plotting the scatter by time series.It could also be character of interval specification,See \code{\link{cut.Date}} for detail.}
}
\value{
a \bold{plotPar} object
}
\description{
set the TSFRScatter related parametres of the ploting
}
\author{
Ruifei.Yin
}
\seealso{
\code{\link{plot.TSFRScatter}}

Other plotPar setting functions: \code{\link{plotPar.IC}},
  \code{\link{plotPar.MC}}, \code{\link{plotPar.Ngroup}},
  \code{\link{plotPar.default}},
  \code{\link{plotPar.longshort}}
}

