% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct02_portBuildingFuncs.R
\name{port.getVolatility}
\alias{port.getVolatility}
\title{GET VOLATILITY OF PORT}
\usage{
port.getVolatility(port)
}
\arguments{
\item{port}{requires colnames: date, stockID and wgt.}
}
\value{
data frame, including two columns: date, port_vol
}
\description{
GET VOLATILITY OF PORT
}
\examples{
port <- getIndexCompWgt("EI000300", endT = as.Date(c("2018-01-31", "2018-03-31")))
port.getVolatility(port)

}
