% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct02_portBuildingFuncs.R
\name{port.wgt_align}
\alias{port.wgt_align}
\title{port.wgt_align}
\usage{
port.wgt_align(port, target_long = 1, target_short = -1)
}
\arguments{
\item{target_long}{long position's target total wgt}

\item{target_short}{short position's target total wgt}
}
\value{
a port
}
\description{
port.wgt_align
}
\examples{
library(data.table)
port <- getIndexCompWgt("EI000300", endT = as.Date(c("2018-01-31", "2018-03-31")))
data.table(port)[,.(sumwgt=sum(wgt)),by=date]
#- wgt aligning
port1 <- port[20:580,]
data.table(port1)[,.(sumwgt=sum(wgt)),by=date]
port1 <- port.wgt_align(port1)
data.table(port1)[,.(sumwgt=sum(wgt)),by=date]
#- deal with the longshort port
port2 <- port
port2[200:400,"wgt"] <- -0.01
port2$tag <- port2$wgt>0
data.table(port2)[,.(sumwgt=sum(wgt)),by=.(date,tag)]
port2 <- port.wgt_align(port2,target_long = 1, target_short = -2)
data.table(port2)[,.(sumwgt=sum(wgt)),by=.(date,tag)]
}
