% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{raw_factor_correlation}
\alias{MC.chart.fCorr}
\alias{MC.table.fCorr}
\alias{raw_factor_correlation}
\title{raw_factor_correlation}
\usage{
MC.chart.fCorr(TSF, Nbin)

MC.table.fCorr(TSF, Nbin)
}
\arguments{
\item{TSF}{is a \bold{TSF} object.}

\item{Nbin}{the number of the groups the timespan is cut to when plotting the scatter by time series.It could also be a character of interval specification,See \code{\link{cut.Date}} for detail. The default value is "day",which means no cutting, the scatters of every date are ploted.}
}
\description{
raw_factor_correlation

MC.table.fCorr
}
\examples{
RebDates <- getRebDates(as.Date('2014-01-31'),as.Date('2016-09-30'))
TS <- getTS(RebDates,indexID = 'EI000985')
factorIDs <- c("F000006","F000008","F000012")
FactorLists <- buildFactorLists_lcfs(factorIDs,factorStd="norm",factorNA = "mean")
TSF <- getMultiFactor(TS,FactorLists = FactorLists)
MC.chart.fCorr(TSF)
MC.chart.fCorr(TSF,Nbin='year')
#-------------------------------------------------------#
MC.table.fCorr(TSF)
MC.table.fCorr(TSF,Nbin='year')
}

