% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{refinePar_default}
\alias{refinePar_default}
\title{refinePar_default}
\usage{
refinePar_default(type = c("none", "robust"),
  sectorAttr = defaultSectorAttr())
}
\arguments{
\item{sectorAttr}{could be a sectorAttr list, or NULL(means no grouping by sector)}
}
\description{
refinePar_default
}
\author{
Han.Qian
}
