% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{reg.factorlists.recommend}
\alias{reg.factorlists.recommend}
\title{factorlists recommend}
\usage{
reg.factorlists.recommend(indexID)
}
\description{
factorlists recommend
}
\examples{
FactorLists <- reg.factorlists.recommend(indexID='EI000300')
FactorLists <- reg.factorlists.recommend(indexID='EI000905')
FactorLists <- reg.factorlists.recommend(indexID='EI000985')
FactorLists <- reg.factorlists.recommend(indexID='ES33370000')
}

