% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{reg.factorlists_recommend}
\alias{reg.factorlists_recommend}
\title{factorlists recommend}
\usage{
reg.factorlists_recommend(indexID, begT, endT, rebFreq = "month",
  rsqBar = 1, forder)
}
\arguments{
\item{indexID}{is index ID.}
}
\description{
factorlists recommend
}
\examples{
##################get the recommended factorLists of last 12 months########## 
begT <- Sys.Date()-lubridate::years(1)
endT <- Sys.Date()-1
indexID <- 'EI000905'
FactorLists <- reg.factorlists_recommend(indexID,begT,endT)
##################get the recommended factorLists of last 4 weeks########## 
begT <- Sys.Date()-months(1)
endT <- Sys.Date()-1
indexID <- 'EI000985'
FactorLists <- reg.factorlists_recommend(indexID,begT,endT,rebFreq = "week")
}
