% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{rm_priceLimit}
\alias{rm_priceLimit}
\title{remove price limits}
\usage{
rm_priceLimit(TS, nearby = 1, lim = c(-10, 10), priceType = c("close",
  "open"), datasrc = defaultDataSRC())
}
\arguments{
\item{lim}{a vector of length 2.}

\item{priceType}{"close" or "open".}
}
\description{
remove price limits
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
re <- rm_priceLimit(TS)
re1 <- rm_priceLimit(TS,nearby=-1:1)
re2 <- rm_priceLimit(TS,lim=c(-Inf,10)) # remove limit-up
}
