% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{rm_suspend}
\alias{rm_suspend}
\title{remove suspension stock from TS}
\usage{
rm_suspend(TS, nearby = 1L, datasrc = defaultDataSRC())
}
\arguments{
\item{TS}{}

\item{nearby}{a integer vector. default is 1L, which means remove the suspending of the next tradingday. see detail in \code{\link{trday.nearby}}.}
}
\description{
remove suspension stock from TS
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
re <- rm_suspend(TS) # remove Suspend of nextday
re1 <- rm_suspend(TS,nearby=c(0,1))# remove Suspend of today and nextday
}

