% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{table.Fct_anova}
\alias{table.Fct_anova}
\alias{chart.Fct_anova}
\title{ANOVA ANALYSIS}
\usage{
table.Fct_anova(TSF, sectorAttr_lists, sectorAttr_names,
  significant_level = 0.05, full_details = FALSE)

chart.Fct_anova(TSF, sectorAttr_lists, sectorAttr_names,
  significant_level = 0.05, value_type = c("p_value", "f_value"))
}
\arguments{
\item{TSF}{A TSF.}

\item{sectorAttr_lists}{A list of sectorAttr, each sectorAttr is a list.}

\item{sectorAttr_names}{A character vector of names, could be missing.}

\item{significant_level}{The ceiling of the p_value. This argument will make sense only when the value_type is p_value. Only the value under this cutting line will be considered as passing the test.}

\item{full_details}{Logical value. Whether return the details instead of summary.}
}
\value{
If all the arguments are default, the result is a table with the ANOVA test pass ratio in each sector splitting method.
}
\description{
ANOVA ANALYSIS
}
\examples{
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.NP_YOY(TS, src = "fin")
sectorAttr_lists_1 <- list(list(std = 33, level = 1),
                           list(std = 336, level = 1))
sectorAttr_lists_2 <- list(list(std = 33, level = 1))
chart.Fct_anova(TSF, sectorAttr_lists_1)
chart.Fct_anova(TSF, sectorAttr_lists_2)
table.Fct_anova(TSF, sectorAttr_lists_1)
table.Fct_anova(TSF, sectorAttr_lists_2)
table.Fct_anova(TSF, sectorAttr_lists_1, full_details = TRUE)
table.Fct_anova(TSF, sectorAttr_lists_2, full_details = TRUE)
}
\author{
Han.Qian
}
