% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{table.ICandNgroup}
\alias{table.ICandNgroup}
\alias{MC.table.ICandNgroup}
\title{table.ICandNgroup}
\usage{
table.ICandNgroup(TSFR, stat = c("pearson", "spearman"), N = 5,
  sectorNe = NULL, fee = 0, rtn_type = c("long-short", "long-univ"))

MC.table.ICandNgroup(TSFRs, stat = c("pearson", "spearman"), N = 5,
  sectorNe = NULL, fee = 0, rtn_type = c("long-short", "long-univ"))
}
\description{
table.ICandNgroup
}
