% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{tables.PA}
\alias{tables.PA}
\title{getPAData}
\usage{
tables.PA(port, factorLists, bmk, sectorAttr = list(33, 1))
}
\description{
getPAData
}
\examples{
alphaLists <- buildFactorLists_lcfs(c("F000012","F000008"),factorStd="norm",factorNA = "median")
riskLists <- buildFactorLists_lcfs(c("F000002","F000006"),,factorStd="norm",factorNA = "median")
re <- getPAData(port,c(alphaLists,riskLists))
}

