% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{tables.PB}
\alias{tables.PB}
\title{tables.PB}
\usage{
tables.PB(PB, hitFreq = "month")
}
\arguments{
\item{PB}{a PB object or a one colume rtn series.}
}
\value{
a list containing some tables which giving the summary result of the PB.
}
\description{
tables.PB
}
\seealso{
\code{\link{tables.longshort}}
}
