% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{ts}
\alias{ts}
\title{demo dataset: TSF.}
\usage{
ts(data = NA, start = 1, end = numeric(), frequency = 1, deltat = 1,
  ts.eps = getOption("ts.eps"), class = if (nseries > 1) c("mts", "ts",
  "matrix") else "ts", names = if (!is.null(dimnames(data))) colnames(data)
  else paste("Series", seq(nseries)))
}
\description{
a TSF object.
}
\keyword{datasets}
