% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{update.lcdb.RegTables}
\alias{update.lcdb.RegTables}
\title{update local database's regression tables}
\usage{
\method{update}{lcdb.RegTables}(FactorLists)
}
\description{
update local database's regression tables
}
\examples{
build.lcdb.RegTables(FactorLists)
}

