% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{wgtSetDemo}
\alias{wgtSetDemo}
\title{WgtSetDemo}
\usage{
wgtSetDemo(secIn = FALSE, stockIn = FALSE)
}
\description{
weight setting demo in optimization function
}
\examples{
wgt <- wgtSetDemo()
wgt <- wgtSetDemo(secIn = T)
wgt <- wgtSetDemo(stockIn = T)
}

