% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{backtest.IC}
\alias{MC.chart.IC}
\alias{MC.chart.IC.corr}
\alias{MC.chart.IC.decay}
\alias{MC.table.IC}
\alias{backtest.IC}
\alias{chart.IC}
\alias{chart.IC.decay}
\alias{seri.IC}
\alias{seri.IC.decay}
\alias{table.IC}
\alias{table.IC.decay}
\title{backtest.IC}
\usage{
seri.IC(TSFR, stat = c("pearson", "spearman"), backtestPar)

seri.IC.decay(TSFR, stat = c("pearson", "spearman"), backtestPar)

table.IC(TSFR, stat = c("pearson", "spearman"), backtestPar)

table.IC.decay(TSFR, stat = c("pearson", "spearman"), backtestPar)

chart.IC(TSFR, Nbin = "day", stat = c("pearson", "spearman"), plotPar)

chart.IC.decay(TSFR, stat = c("pearson", "spearman"), plotPar)

MC.chart.IC.corr(TSFRs, stat = c("pearson", "spearman"), plotPar)

MC.table.IC(TSFRs, stat = c("pearson", "spearman"), backtestPar)

MC.chart.IC(TSFRs, Nbin = "day", stat = c("pearson", "spearman"),
  ncol = 3, plotPar)

MC.chart.IC.decay(TSFRs, stat = c("pearson", "spearman"), ncol = 3, plotPar)
}
\arguments{
\item{TSFR}{a \bold{TSFR} object}

\item{stat}{a character string,indicating the methods to compute IC,could be "pearson" or "spearman".}

\item{backtestPar}{Optional.a \bold{backtestPar} object,if not missing,then extract pars from backtestPar.}

\item{Nbin}{the number of the groups the timespan is cut to, when plotting the IC series.It could also be character of interval specification,See \code{\link{cut.Date}} for detail. The default value is "day",which means no cutting, the value of every date are ploted.}

\item{plotPar}{Optional.a \bold{plotPar} object,if not missing,then extract pars from plotPar}

\item{TSFRs}{a list of object \bold{TSFR}. See \code{\link{Model.TSFR}}.}

\item{ncol}{a integer, specificate the number of cols of the multi-charts.}
}
\value{
seri.IC return a xts object, which containing the IC seri

seri.IC.decay return a xts object of 12 cols, which containing the decayed ICs seri

table.IC return a matirx of the statistical of the IC, containing rows: "IC.mean","IC.std","IC.IR","IC.t","IC.p","IC.hitRatio"

table.IC.decay return a matirx of the statistical of the decayed ICs, containing 5 rows: "IC.mean","IC.std","IC.IR","IC.t","IC.p","IC.hitRatio" and 12 cols: "IC1"-"IC12".

chart.IC return a ggplot object of IC time series(with its 12 months MA)

chart.IC.decay return a ggplot object of decayed ICs bar chart.

MC.chart.IC.corr return a correlation chart of ICs of each \code{TSFR}.
}
\description{
backtesting the factor with some tables and charts using the 'IC' method.
}
\details{
When caculating the correlation,two methods "pearson" and "spearman" is used.

If param backtestPar and plotPar is not missing,then the related params will be extracted from them.It is usefull when the parametres has been initialized.
}
\examples{
modelPar <- modelPar.default()
TSFR <- Model.TSFR(modelPar)
re <- seri.IC(TSFR)
re <- seri.IC.decay(TSFR)
IC.table <- table.IC(TSFR)
IC.table.decay <- table.IC.decay(TSFR)
IC.chart <- chart.IC(TSFR,"3 month")
re <- chart.IC.decay(TSFR)
mp = modelPar.default()
factorIDs <- c("F000001","F000002","F000005")
FactorLists <- buildFactorLists_lcfs(factorIDs)
mps <- getMPs_FactorLists(FactorLists,modelPar=mp)
TSR <- Model.TSR(mp)
TSFRs <- Model.TSFs_byTS(MPs=mps,TS=TSR)
MC.chart.IC.corr(TSFRs)
MC.table.IC(TSFRs)
MC.chart.IC(TSFRs)
MC.chart.IC.decay(TSFRs)
}
\author{
Ruifei.Yin
}

