% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{backtest.longshort}
\alias{backtest.longshort}
\alias{tables.longshort}
\alias{chart.longshort.summary}
\alias{chart.longshort.rolling}
\title{backtest.longshort}
\usage{
tables.longshort(rtn.LSH, hitFreq = "month")

chart.longshort.summary(rtn.LSH, bar.freq = "month")

chart.longshort.rolling(rtn.LSH, roll.width = 250, roll.by = 30)
}
\arguments{
\item{rtn.LSH}{a \bold{rtn.LSH} or a \bold{rtn.LBH} object getting by function \code{\link{getrtn.LSH}} or \code{\link{getrtn.LBH}}.}

\item{hitFreq}{indicating the interval when computing the hitRatio of rtn. An interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s".See \code{\link{cut.Date}} for detail.}

\item{bar.freq}{the freq of the per-period performance bar chart}

\item{roll.width}{the width argument for rolling performance chart}

\item{roll.by}{the by argument for rolling performance chart}
}
\value{
tables.longshort return a list containing some tables which giving the result of the long-short strategy backtesting.The items are:
 \itemize{
   \item summary: summary of \bold{of long,short and hedge}.
   \item summary.yearly: yearly summary \bold{of the hedged rtn}.
   \item hedge.stats: main statisticals \bold{of the hedged rtn},by different freq.
   \item period.stats: table showing yearly,all-span and annualized return \bold{of long,short and hedge}.
   \item DD.stats:table showing statistics for the worst drawdowns \bold{of the hedged rtn}.
 }

chart.longshort.summary  return and print a recordedplot object, which demonstrate the performance of the return series,including wealth index chart(\bold{of long,short and hedging}),underwater chart for drawdown(\bold{of hedging}),and bars for per-period performance(\bold{of hedging})..

chart.longshort.rolling  return and print a recordedplot object, which include a rolling annualized returns chart,a rolling annualized standard deviation chart, and a rolling annualized sharpe ratio chart.
}
\description{
backtesting the factor with some tables and charts using the 'long-short(hedging)' method.
}
\examples{
rtn.long <- xts(rnorm(1000,0.001,0.02),as.Date("2010-01-01") + 1:1000)
rtn.short <- rtn.long + rnorm(1000,-0.0001,0.003)
rebFreq <- "month"
rtn.LSH <- addrtn.hedge(rtn.long,rtn.short,rebFreq)
re <- tables.longshort(rtn.LSH)
chart.longshort.summary(rtn.LSH)
chart.longshort.rolling(rtn.LSH)
}
\seealso{
\code{\link{getrtn.LSH}},\code{\link{getrtn.LBH}}
}
\author{
Ruifei.Yin
}
