% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{buildWgtSet}
\alias{buildWgtSet}
\title{buildWgtSet}
\usage{
buildWgtSet(wgtall = c(0, 0.01), ...)
}
\description{
build weight setting in optimization function
}
\examples{
wgtSet <- buildWgtSet()
wgtSet <- buildWgtSet(wgtall=c(0,0.02),EQ601318=c(0.05,0.15))
wgtSet <- buildWgtSet(EQ601318=c(0.05,0.15),ES33370000=c(0,0.5))
}
