% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{f_rtn_cov}
\alias{f_rtn_cov}
\alias{getfRtn}
\alias{calfCov}
\title{factor return and covariance}
\usage{
getfRtn(RebDates, fNames, dure = months(1), type = c("mean", "rollmean",
  "forcast"), nwin = lubridate::years(-2), reg_results)

calfCov(RebDates, fNames, dure = months(1), covtype = c("robust", "simple",
  "roll-simple", "roll-robust"), nwin, reg_results)
}
\arguments{
\item{dure}{a period object from package \code{lubridate}. (ie. \code{months(1),weeks(2)}. See example in \code{\link{trday.offset}}.) If null, then get periodrtn between \code{date} and the next \code{date}, else get periodrtn of '\code{dure}' starting from \code{date}.}

\item{type}{}

\item{nwin}{rolling windows.}
}
\value{
a factor return data frame.
}
\description{
calculate factor return and factor covariance.
}
\examples{
RebDates <- getRebDates(as.Date('2014-01-31'),as.Date('2016-08-31'))
fNames <- c("NP_YOY","PB_mrq_","disposition_","ln_mkt_cap_")
fRtn <- getfRtn(RebDates,fNames,reg_results=reg_results)
fCov <- calfCov(RebDates,fNames,reg_results=reg_results)
}
