% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{factor.VIF}
\alias{factor.VIF}
\title{factor.VIF}
\usage{
factor.VIF(TSF, testf, sectorAttr = defaultSectorAttr())
}
\arguments{
\item{TSF}{is a \bold{TSF} object.}

\item{testf}{is test factor name.}

\item{sectorAttr}{is sector attribute.}
}
\value{
data frame of VIF or residual.
}
\description{
get factor's VIF or stock's residual
}
\examples{
VIF <- factor.VIF(TSF)
VIF <- factor.VIF(TSF,testf)[[1]]
res <- factor.VIF(TSF,testf)[[2]]
}
\author{
Andrew Dow
}

