% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{factor.VIF}
\alias{factor.VIF}
\title{get factor's VIF}
\usage{
factor.VIF(TS, testfactorList, factorLists, sectorAttr = defaultSectorAttr())
}
\arguments{
\item{TS}{is a \bold{TS} object.}

\item{testfactorList}{is a factor list for test.}

\item{factorLists}{is a set of factors in use.}

\item{sectorAttr}{is sector attribution.}
}
\value{
a VIF time series.
}
\description{
get factor's VIF
}
\examples{
RebDates <- getRebDates(as.Date('2012-12-31'),as.Date('2016-08-31'),rebFreq = 'month')
TS <- getTS(RebDates,'EI000985')
testfactorList <- buildFactorLists_lcfs('F000002',factorStd="norm",factorNA = "median")
factorIDs <- c("F000003","F000004","F000007","F000008","F000011")
factorLists <- buildFactorLists_lcfs(factorIDs,factorStd="norm",factorNA = "median")
VIF <- factor.VIF(TS,testfactorList,factorLists)
}
\author{
Andrew Dow
}

