% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{factor_descriptive_statistics}
\alias{factor_descriptive_statistics}
\alias{chart.TSF_hist}
\alias{chart.TSF_density}
\alias{chart.TSF_box}
\alias{table.TSF_descr}
\alias{MC.chart.TSF_hist}
\alias{MC.chart.TSF_density}
\alias{MC.chart.TSF_box}
\alias{MC.table.TSF_descr}
\title{Factor descriptive statistics}
\usage{
chart.TSF_hist(TSF, Nbin = "day")

chart.TSF_density(TSF, Nbin = "day")

chart.TSF_box(TSF, Nbin = "day")

table.TSF_descr(TSF, Nbin = "day")

MC.chart.TSF_hist(TSF)

MC.chart.TSF_density(TSF, Nbin = "day")

MC.chart.TSF_box(TSF, Nbin = "day")

MC.table.TSF_descr(TSF)
}
\description{
draw factor's histogram and boxplot,and summarize factor's statistics value.
}
\examples{
mp <- modelPar.default()
TSF <- Model.TSF(mp)
chart.TSF_hist(TSF) #plot each period
chart.TSF_hist(TSF,Nbin='year') #plot each year
chart.TSF_box(TSF)
chart.TSF_density(TSF)
re <- table.TSF_descr(TSF)
####################################~~ multiple factor ~~##################
FactorLists <- buildFactorLists_lcfs(c("F000006","F000002","F000005"))
TSF <- getMultiFactor(TSF[,c('date','stockID')],FactorLists)
MC.chart.TSF_hist(TSF)
MC.chart.TSF_box(TSF)
MC.chart.TSF_density(TSF)
re <- MC.table.TSF_descr(TSF)
}
