% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{factor_descriptive_statistics}
\alias{factor_descriptive_statistics}
\alias{chart.Fct_hist}
\alias{chart.Fct_density}
\alias{chart.Fct_box}
\alias{table.Fct_descr}
\alias{MF.chart.Fct_hist}
\alias{MF.chart.Fct_density}
\alias{MF.chart.Fct_box}
\alias{MF.table.Fct_descr}
\alias{factor_descriptive_statistics}
\alias{MF.chart.Fct_corr}
\alias{MF.table.Fct_corr}
\title{Factor descriptive statistics}
\usage{
chart.Fct_hist(TSF, bins = NULL, ncol = NULL)

chart.Fct_density(TSF)

chart.Fct_box(TSF)

table.Fct_descr(TSF)

MF.chart.Fct_hist(mTSF)

MF.chart.Fct_density(mTSF, ncol = NULL)

MF.chart.Fct_box(mTSF, ncol = NULL)

MF.table.Fct_descr(mTSF)

MF.chart.Fct_corr(mTSF, Nbin)

MF.table.Fct_corr(mTSF, Nbin)
}
\arguments{
\item{Nbin}{the number of the groups the timespan is cut to when plotting the scatter by time series.It could also be a character of interval specification,See \code{\link{cut.Date}} for detail. The default value is "day",which means no cutting, the scatters of every date are ploted.}
}
\description{
draw factor's histogram and boxplot,and summarize factor's statistics value.
}
\examples{
mp <- modelPar.default()
TSF <- Model.TSF(mp)
chart.Fct_hist(TSF)
chart.Fct_box(TSF)
chart.Fct_density(TSF)
re <- table.Fct_descr(TSF)
#~~ multiple factor ~~
FactorLists <- buildFactorLists_lcfs(c("F000006","F000002","F000005"))
TS <- Model.TS(mp)
mTSF <- getMultiFactor(TS,FactorLists)
MF.chart.Fct_hist(mTSF)
MF.chart.Fct_box(mTSF)
MF.chart.Fct_density(mTSF)
re2 <- MF.table.Fct_descr(mTSF)
# ---  raw_factor_correlation
RebDates <- getRebDates(as.Date('2014-01-31'),as.Date('2016-09-30'))
TS <- getTS(RebDates,indexID = 'EI000985')
factorIDs <- c("F000006","F000008","F000012")
FactorLists <- buildFactorLists_lcfs(factorIDs,factorRefine=refinePar_default("robust"))
mTSF <- getMultiFactor(TS,FactorLists = FactorLists)
MF.chart.Fct_corr(mTSF)
MF.chart.Fct_corr(mTSF,Nbin='year')
MF.table.Fct_corr(mTSF)
MF.table.Fct_corr(mTSF,Nbin='year')
}
\author{
Aming.Tao
}
