% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{getfRtn}
\alias{getfRtn}
\title{getfRtn}
\usage{
getfRtn(RebDates, fNames, dure = months(1), rollavg = TRUE,
  nwin = lubridate::years(-2), reg_results)
}
\arguments{
\item{dure}{a period object from package \code{lubridate}. (ie. \code{months(1),weeks(2)}. See example in \code{\link{trday.offset}}.) If null, then get periodrtn between \code{date} and the next \code{date}, else get periodrtn of '\code{dure}' starting from \code{date}.}

\item{rollavg}{whether to get the rolling average factor return.}

\item{nwin}{rolling windows.}
}
\value{
a factor return data frame.
}
\description{
get factor return data
}
\examples{
RebDates <- getRebDates(as.Date('2014-01-31'),as.Date('2016-08-31'))
fNames <- c("ln_mkt_cap_","PB_mrq_","pct_chg_per_60_","IVR_")
dure <- lubridate::days(1)
}

