% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{getPAData}
\alias{getPAData}
\title{getPAData}
\usage{
getPAData(port, factorLists, bmk, sectorAttr = defaultSectorAttr())
}
\description{
getPAData
}
\examples{
tmp <- buildFactorLists(buildFactorList(factorFun="gf.NP_YOY",
                factorPar=list(),factorDir=1),factorStd="norm",factorNA = "median")
alphaLists <- buildFactorLists_lcfs(c("F000012","F000008"),factorStd="norm",factorNA = "median")
alphaLists <- c(tmp,alphaLists)
riskLists <- buildFactorLists_lcfs(c("F000002","F000006"),factorStd="norm",factorNA = "median")
PA_tables <- getPAData(port,c(alphaLists,riskLists))
PA_tables <- getPAData(port,c(alphaLists,riskLists),bmk='EI000905')
}
