% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{getPAData}
\alias{getPAData}
\title{getPAData}
\usage{
getPAData(port, factorLists, bmk = NULL, univ = "EI000985",
  sectorAttr = defaultSectorAttr())
}
\description{
getPAData
}
\examples{
alphaLists <- buildFactorLists_lcfs(c("F000012","F000008"),factorRefine=refinePar_default("scale"))
riskLists <- buildFactorLists_lcfs(c("F000002","F000006"),factorRefine=refinePar_default("none"))
PA_tables <- getPAData(port,c(alphaLists,riskLists))
PA_tables <- getPAData(port,c(alphaLists,riskLists),bmk='EI000905',univ="EI000905")
}
