% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct02_portBuildingFuncs.R
\name{getrtn.bmk}
\alias{getrtn.bmk}
\title{getrtn.bmk}
\usage{
getrtn.bmk(rtn, bmk, date_start_pad)
}
\arguments{
\item{rtn}{an xts,timeSeries or zoo object of portfolio returns.}

\item{bmk}{a character string, giving the stockID of the benchmark index,eg. "EI000300".}

\item{date_start_pad}{a Date value, indicating the start point of the first rtn period. If missing, padded by an approximate value.}
}
\value{
a xts object, giving the rtn series of the benchmark
}
\description{
get the benchmark return series corresponding to the specific rtn series.
}
\examples{
# -- daily rtn
dates <- getRebDates(as.Date("2011-03-01"),as.Date("2011-04-20"),rebFreq = "day")
TS <- getTS(dates,stocks = "EQ000002")
TSR <- getTSR(TS)
rtn <- xts::xts(TSR$periodrtn,TSR$date_end)
re <- getrtn.bmk(rtn,"EI000300")
# -- IF00 as bmk
re_IF <- getrtn.bmk(rtn,"IF00")
# -- none daily rtn
dates <- getRebDates(as.Date("2010-03-01"),as.Date("2011-04-20"),rebFreq = "month")
TS <- getTS(dates,stocks = "EQ000002")
TSR <- getTSR(TS)
rtn <- xts::xts(TSR$periodrtn,TSR$date_end)
re <- getrtn.bmk(rtn,"EI000300")
re2 <- getrtn.bmk(rtn,"EI000300",date_start_pad=as.Date("2010-03-01"))
}
\seealso{
Other LSH-frame building functions: \code{\link{addrtn.hedge}},
  \code{\link{getrtn.LBH}}, \code{\link{getrtn.LSH}}
}
\author{
Ruifei.Yin
}
