% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct02_portBuildingFuncs.R
\name{getrtn.LSH}
\alias{getrtn.LSH}
\title{getrtn.LSH}
\usage{
getrtn.LSH(PB.L, PB.S, hedge.rebFreq = "month", hedge.posi = 1,
  fee.long = 0, fee.short = 0)
}
\arguments{
\item{PB.L}{a \bold{PB} object of the long portfolio}

\item{PB.S}{a \bold{PB} object of the short portfolio}

\item{hedge.rebFreq}{giving the rebalance freq when computing the hedged rtn.An interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s".See \code{\link{cut.Date}} for detail.}

\item{hedge.posi}{a numeric, giving the position of the hedging assets}

\item{fee.long}{a numeric, giving the fee of long asset}

\item{fee.short}{a numeric, giving the fee of short asset}
}
\value{
a \bold{rtn.LSH} object of class xts,giving the return series of long,short and hedge, with the attr of 'turnover'(a xts series) and 'fee'(a vector).
}
\description{
get the \bold{rtn.LSH} object from the \bold{PB} object.
}
\examples{
port.L <- getPort(TSF,20, dir="long")
port.S <- getPort(TSF,20, dir="short")
PB.L <- port.backtest(port.L)
PB.S <- port.backtest(port.S)
re <- getrtn.LSH(PB.L,PB.S)
}
\seealso{
Other LSH-frame building functions: \code{\link{addrtn.hedge}},
  \code{\link{getrtn.LBH}}, \code{\link{getrtn.bmk}}
}
\author{
Ruifei.Yin
}
