% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{getTSF}
\alias{getMultiFactor}
\alias{getMultiFactorbyTSFs}
\alias{getRawFactor}
\alias{getTSF}
\title{getTSF}
\usage{
getTSF(TS, factorFun, factorPar = list(), factorDir = 1,
  factorOutlier = 3, factorStd = c("sectorNe", "none", "norm"),
  factorNA = c("na", "mean", "median", "min", "max"),
  sectorAttr = defaultSectorAttr())

getRawFactor(TS, factorFun, factorPar)

getMultiFactor(TS, FactorLists, wgts)

getMultiFactorbyTSFs(TSFs, wgts)
}
\arguments{
\item{TS}{a \bold{TS} or \bold{TSR} objectobject(See detail in \code{\link{getTS}} and  \code{\link{getTSR}})}

\item{factorFun}{a non-empty character string naming the function to get the factor scores.}

\item{factorPar}{either a list or a character string, containing the parameters of the \code{factorFun}. If a character string, parameters are seprated by commas and the character parameters must quoted by  "\""s, see examples for details.}

\item{factorDir}{a integer,should be 1 or -1 (1 for the positive factor,-1 for the negative one).}

\item{factorOutlier}{a numeric value of standard deviations, e.g. 10 means that factor score larger or smaller than ten times the standard deviation will be removed}

\item{factorStd}{a character string, indicating the standardizing pattern of the factor score,could be one of "none"(no scaling),"norm"(scaling to 0 mean and 1 stdev),"sectorNe"(scaling to 0 mean and 1 stdev in each sector).}

\item{factorNA}{a character string, indicating the method to deal with the missing value of the factorscore, could be one of "na"(default,keeping the missing value as it is),"mean","median","max","min".}

\item{FactorLists}{a list, with elements of FactorList(See detail in \code{\link{buildFactorList}}).}

\item{wgts}{a numeric vector with the same length of FactorLists and sum of 1(if not, the wgts will be rescaled to sum 1).}

\item{TSFs}{a \bold{TSFs} object. see /code{/link{Model.TSFs}}}
}
\value{
given a \bold{TS} object,return a \bold{TSF} object,a dataframe containing cols:
  \itemize{
  \item date: the rebalance dates, with \code{Date} class
  \item stockID: the stockID contained in the index or BK
  \item factorscore:the factor score of the stocks on the  rebalancing date
  \item sector(optional):the sector of a stock
  }
  given a \bold{TSR} object,return a \bold{TSFR} object,a dataframe containing cols:
  \itemize{ 
  \item stockID: the stockID contained in the index or BK
  \item date: the rebalance dates, with \code{Date} class
  \item factorscore:the factor score of the stocks on the  rebalancing date  
  \item date_end(optional): the next rebalance dates, with \code{Date} class
  \item periodrtn:the period rtn between the rebalance date and the date_end
  \item sector(optional):the sector of a stock
  }

\code{getRawFactor} return a \bold{TSF} object, with 'raw' factorscore.

\code{getMultiFactor} return a \bold{TSF} object including cols of all the single-factor-scores, and (if param \code{wgts} not missing) the multi-factor-score .

\code{getMultiFactorbyTSFs} return a \bold{TSF} object including cols of all the single-factor-scores, and (if param \code{wgts} not missing) the multi-factor-score, from a \bold{TSFs} list. (If param TSFs is a TSFRs object, then will return a TSFR object.)
}
\description{
get the \bold{TSF} ('time-stock-factor') or \bold{TSFR}('time-stock-factor-rtn') object,by adding the factor score to a \bold{TS} or \bold{TSR} object.If necessary,cleaning and stardizing the factor score.
}
\details{
\code{getRawFactor} return a \bold{TSF} object, with 'raw' factorscore, which has not been standardized and not been deal with missing values and outliers. Function \code{getRawFactor} is often called intermediately by function \code{getTSF}.

Function \code{getMultiFactor} get the multi-factor-score by firstly getting all the single-factor-scores and then calculating the weighted sum of them.
}
\note{
Function \code{getMultiFactor} not only could be used to get the "raw" multi-factor-score, but also could be used as the \code{"factorFun"} parametre in function \code{getTSF} to get the "clean and standardized" multi-factor-score. See more detail in the examples.
}
\examples{
# -- get the TSF step by step --

# - get a TSF object
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
factorFun <- "gf.demo"
factorPar <- list(mean=0,sd=1)
TSF <- getTSF(TS,factorFun,factorPar)
# - get a TSFR object
TSR <- getTSR(TS)
TSFR <- getTSF(TSR)

# -- you can also get the TSF through the modelPar object directively --
modelPar <- modelPar.default()
modelPar <- modelPar.factor(modelPar,factorFun="gf.demo",factorPar=list(0,1),factorOutlier=10,factorStd="sectorNe")
TSF <- Model.TSF(modelPar)

# -- factorPar as character string
TSF2 <- getTSF(TS,"gf.demo","0,1")
TSF2 <- getTSF(TS,"gf.demo","mean=0,sd=1")
TSF2 <- getTSF(TS,"gf.foo","2,3")
TSF2 <- getTSF(TS,"gf.bar","20,\\"IF00\\"")
# -- get 'raw' factorscore
TSF_raw <- getRawFactor(TS,"gf.pct_chg_per","20")
# -- get multi-factor by FactorLists 
FactorLists <- list(
  buildFactorList(factorFun="gf.PE",
                  factorPar=list(),
                  factorDir=-1),
  buildFactorList(factorFun="gf.pct_chg_per",
                  factorPar=list(N=20),
                  factorDir=-1),
  buildFactorList(factorFun="gf.pct_chg_per",
                  factorPar=list(N=60),
                  factorDir=-1),       
  buildFactorList(factorFun="gf.pct_chg_per",
                  factorPar=list(N=120),
                  factorDir=-1))
wgts <- c(0.25,0.25,0.25,0.25)

# -- 0. get the all the single-factor-scores
TSF <- getMultiFactor(TS,FactorLists)

# -- 1. get the "raw" multi-factor-score
TSF.multi <- getMultiFactor(TS,FactorLists,wgts)

# -- 2. get the "clean and standardized" multi-factor-score
# --- 2.1 get the multi-factor-score by \\code{getTSF}
TSF.multi2 <- getTSF(TS,factorFun="getMultiFactor",factorPar=list(FactorLists,wgts),factorOutlier=10,factorStd="sectorNe")
# --- 2.2 get the multi-factor-score by \\code{Model.TSF}
modelPar <- modelPar.factor(modelPar.default(),factorFun="getMultiFactor",factorPar=list(FactorLists,wgts),factorStd="sectorNe",factorDir=1)
TSF.multi3 <- Model.TSF(modelPar)
TSF.multi4 <- Model.TSF_byTS(modelPar, TS)
# -- get multi-factor by TSFs 
MPs <- getMPs_FactorLists(FactorLists, modelPar.default())
TSFs <- Model.TSFs(MPs)
TSF.multi5 <- getMultiFactorbyTSFs(TSFs,wgts)
}
\author{
Ruifei.Yin
}
\seealso{
Other frame building functions: \code{\link{getRebDates}},
  \code{\link{getTSR}}, \code{\link{getTS}}
}

