% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{lcdb_regtables}
\alias{lcdb_regtables}
\alias{lcdb.build.RegTables}
\alias{lcdb.update.RegTables}
\title{lcdb_regtables}
\usage{
lcdb.build.RegTables(begT, endT, FactorLists)

lcdb.update.RegTables(begT, endT, FactorLists)
}
\arguments{
\item{begT}{is begin date}

\item{endT}{is end date}

\item{FactorLists}{see example in \code{\link{buildFactorLists}}.}
}
\description{
build and update local database's regression result tables
}
\examples{
begT <- as.Date('2005-01-04')
endT <- as.Date('2009-12-31')
FactorLists <- buildFactorLists(
  buildFactorList(factorFun="gf.SIZE"),
  buildFactorList(factorFun="gf.GROWTH"),
  buildFactorList(factorFun="gf.TRADING"),
  buildFactorList(factorFun="gf.EARNINGYIELD"),
  buildFactorList(factorFun="gf.VALUE"),
  buildFactorList(factorFun="gf.OTHER"))
lcdb.build.RegTables(begT,endT,FactorLists)
begT <- as.Date('2010-01-04')
endT <- as.Date('2014-12-31')
lcdb.update.RegTables(begT,endT,FactorLists)
}
