% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct03_backtestingFuncs.R
\name{MC.wgt.CAPM}
\alias{MC.wgt.CAPM}
\title{MC.wgt.CAPM}
\usage{
MC.wgt.CAPM(TSFRs, stat = c("pearson", "spearman"), backtestPar, wgtmin = 0,
  wgtmax = 0.5, targetType = c("sharpe", "return", "risk", "balance"),
  riskaversion = 1, reg_results)
}
\arguments{
\item{TSFRs}{a list of object \bold{TSFR}. See \code{\link{Model.TSFR}}.}

\item{stat}{a character string,indicating the methods to compute IC,could be "pearson" or "spearman".}

\item{backtestPar}{Optional.a \bold{backtestPar} object,if not missing,then extract pars from backtestPar.}

\item{wgtmin}{set minimal factor weight.}

\item{wgtmax}{set maximal factor weight.}

\item{targetType}{optimization's target type, could be "return" or "risk" or "sharpe" or 'balance',default value is "sharpe".}

\item{riskaversion}{risk aversion parameter for "balance" target.}

\item{reg_results}{See \code{\link{reg.TSFR}}.}
}
\value{
a factor weight vector
}
\description{
compute the wgt vector of multi-factors by CAPM model.
}
\examples{
mp <- modelPar.default()
factorIDs <- c("F000001","F000004","F000005","F000008")
FactorLists <- buildFactorLists_lcfs(factorIDs)
mps <- getMPs_FactorLists(FactorLists,modelPar=mp)
TSR <- Model.TSR(mp)
TSFRs <- Model.TSFs_byTS(MPs=mps,TS=TSR)
MC.wgt.CAPM(TSFRs)
MC.wgt.CAPM(TSFRs,wgtmin=0.05,wgtmax=0.4,targetType='risk')
MC.wgt.CAPM(TSFRs,wgtmin=0.05,wgtmax=0.4,targetType='balance',riskaversion = 10)
-----------------------------------------------------------------------------
MC.wgt.CAPM(reg_results=reg_results)
MC.wgt.CAPM(wgtmin=0.05,wgtmax=0.4,targetType='balance',reg_results=reg_results)
}
