% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct06_portOptimizing.R
\name{opt_constrain}
\alias{opt_constrain}
\alias{constr_default}
\alias{clearConstr}
\alias{addConstr_fctExp_sector}
\alias{addConstr_fctExp_style}
\alias{addConstr_group}
\alias{addConstr_box}
\alias{addConstr_position}
\alias{addConstr_turnover}
\alias{addConstr_trackingerror}
\alias{setConstr_group}
\alias{setConstr_box}
\alias{setConstr_fctExp_sector}
\alias{setConstr_position}
\alias{setConstr_fctExp_style}
\alias{setConstr_turnover}
\alias{setConstr_trackingerror}
\title{opt_constrain}
\usage{
constr_default(position = c(1, 1), box_each = c(0, 0.1))

clearConstr(constr, item)

addConstr_fctExp_sector(constr, each, ..., relative = 2,
  sectorAttr = defaultSectorAttr())

addConstr_fctExp_style(constr, FactorLists, min, max, relative = 1)

addConstr_group(constr, ..., relative = 0)

addConstr_box(constr, each, ..., relative = 0)

addConstr_position(constr, position, relative = 0)

addConstr_turnover(constr, turnover_target = 0.25, method = c("rmosek",
  "matlab"))

addConstr_trackingerror(constr, trackingerror_ann = 0.05,
  method = c("rmosek", "matlab"))

setConstr_group(constr, ..., relative)

setConstr_box(constr, each, ..., relative)

setConstr_fctExp_sector(constr, each, ..., relative,
  sectorAttr = defaultSectorAttr())

setConstr_position(constr, position, relative,
  sectorAttr = defaultSectorAttr())

setConstr_fctExp_style(constr, FactorLists, min, max, relative)

setConstr_turnover(constr, turnover_target, ...)

setConstr_trackingerror(constr, trackingerror_ann, ...)
}
\arguments{
\item{relative}{integer of 0,1,2}

\item{FactorLists}{a factorlists}

\item{min}{a vector with the same length of FactorLists, or a scalar}

\item{max}{a vector with the same length of FactorLists, or a scalar}
}
\value{
a list contain the constrains
}
\description{
build, set, add, clear constrains
}
\examples{
constr <- constr_default()
clearConstr(constr,"box")
addConstr_fctExp_sector(constr, each=c(-0.05,0.05),ES33480000=c(-0.15,0.15))
factorlists <- buildFactorLists_lcfs(factorIDs = c("F000001","F000002"),factorRefine=refinePar_default("scale",NULL))
addConstr_fctExp_style(constr,factorlists,-0.1,0.1)
addConstr_group(constr,EI000905=c(0, 0.8))
addConstr_group(constr,EI000001=c(0.5, 0.5))
addConstr_group(constr,ES33220000=c(-0.1, 0.1),relative = 1)
addConstr_box(constr,each=c(0,0.02),EQ601318=c(0.05,0.15))
addConstr_box(constr,EQ601318=c(0.05,0.15),ES33370000=c(0,0.5))
addConstr_position(constr,position=c(0.8,1))
addConstr_turnover(constr,turnover_target=0.5)
addConstr_trackingerror(constr,trackingerror_ann=0.08)
}
