% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct06_portOptimizing.R
\name{opt_constrain}
\alias{opt_constrain}
\alias{constr_default}
\alias{clearConstr}
\alias{addConstr_fctExp_sector}
\alias{addConstr_fctExp_style}
\alias{addConstr_group}
\alias{addConstr_box}
\alias{addConstr_position}
\alias{addConstr_leverage}
\alias{addConstr_turnover}
\alias{addConstr_trackingerror}
\alias{addConstr_sus_lim}
\alias{addConstr_FFV}
\alias{setConstr_group}
\alias{setConstr_box}
\alias{setConstr_fctExp_sector}
\alias{setConstr_position}
\alias{setConstr_leverage}
\alias{setConstr_fctExp_style}
\alias{setConstr_turnover}
\alias{setConstr_trackingerror}
\alias{setConstr_sus_lim}
\alias{setConstr_FFV}
\title{opt_constrain}
\usage{
constr_default(position = c(1, 1), box_each = c(0, 0.1))

clearConstr(constr, item)

addConstr_fctExp_sector(constr, each, ..., relative = 2,
  sectorAttr = defaultSectorAttr())

addConstr_fctExp_style(constr, FactorLists, min, max, relative = 1)

addConstr_group(constr, ..., relative = 0)

addConstr_box(constr, each, ..., relative = 0, priority = 0)

addConstr_position(constr, position, relative = 0)

addConstr_leverage(constr, leverage)

addConstr_turnover(constr, turnover_target = 0.25)

addConstr_trackingerror(constr, trackingerror_ann = 0.05)

addConstr_sus_lim(constr, ID = c("sus", "lim_up", "lim_low"))

addConstr_FFV(constr, TW_ratio = 3)

setConstr_group(constr, ..., relative)

setConstr_box(constr, each, ..., relative, priority)

setConstr_fctExp_sector(constr, each, ..., relative,
  sectorAttr = defaultSectorAttr())

setConstr_position(constr, position, relative)

setConstr_leverage(constr, leverage)

setConstr_fctExp_style(constr, FactorLists, min, max, relative)

setConstr_turnover(constr, turnover_target, ...)

setConstr_trackingerror(constr, trackingerror_ann, ...)

setConstr_sus_lim(constr, ID)

setConstr_FFV(constr, TW_ratio)
}
\arguments{
\item{constr}{a list contained the constraining setting.}

\item{relative}{integer of 0,1,2}

\item{FactorLists}{a factorlists}

\item{min}{a vector with the same length of FactorLists, or a scalar}

\item{max}{a vector with the same length of FactorLists, or a scalar}

\item{priority}{integer(only the max priority will make effect on the same stock), NA(not involved in the integeral priority, enforce to get the intersetion in the final result.)}
}
\value{
a list contain the constrains
}
\description{
build, set, add, clear constrains
}
\examples{
constr <- constr_default()
clearConstr(constr,"box")
addConstr_fctExp_sector(constr, each=c(-0.05,0.05),ES33480000=c(-0.15,0.15))
factorlists <- buildFactorLists_lcfs(factorIDs = c("F000001","F000002"),factorRefine=refinePar_default("scale",NULL))
addConstr_fctExp_style(constr,factorlists,-0.1,0.1)
addConstr_group(constr,EI000905=c(0, 0.8))
addConstr_group(constr,EI000001=c(0.5, 0.5))
addConstr_group(constr,ES33220000=c(-0.1, 0.1),relative = 1)
constr <- constr_default()
constr <- addConstr_box(constr,each=c(0,0.035),relative = 0,priority = NA)
constr <- addConstr_box(constr,EQ002011=c(0,0.05),EQ000028=c(0,0.05),priority = 2)
constr <- addConstr_box(constr,EI399005=c(-0.2,0.2),relative = 1,priority = 1)
constr <- addConstr_box(constr,ES33370000=c(0,0.03),relative = 0,priority = 1)
constr <- addConstr_box(constr,ES33330000=c(-0.1,0.1),relative = 2,priority = 1)
addConstr_position(constr,position=c(0.8,1))
addConstr_leverage(constr,leverage=1.6) #for 130/30
addConstr_turnover(constr,turnover_target=0.5)
addConstr_trackingerror(constr,trackingerror_ann=0.08)
addConstr_sus_lim(constr,ID=c("sus","lim_up"))
addConstr_FFV(constr,TW_ratio = 3)
}
