% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct06_portOptimizing.R
\name{opt_object}
\alias{opt_object}
\alias{object_default}
\alias{clearObj}
\alias{addObj_return}
\alias{addObj_risk}
\title{opt_object}
\usage{
object_default()

clearObj(obj, item)

addObj_return(obj, method = c("mean", "event"))

addObj_risk(obj, method = c("rmosek", "solve.QP"), risk_aversion = 4)
}
\value{
a list contain the objects
}
\description{
build, set, add, clear objects
}
\examples{
obj <- object_default()
obj <- addObj_risk(obj)
clearObj(obj,"return")
}
