% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{OptWgt}
\alias{OptWgt}
\title{OptWgt}
\usage{
OptWgt(TSF, alphaf, fRtn, fCov, target = c("return", "return-risk"),
  constr = c("IndSty", "Ind", "IndStyTE"), benchmark = "EI000905",
  indfexp = 0.05, fexp, sectorAttr = defaultSectorAttr(), maxWgt = 0.01,
  addEvent = FALSE, optWay = c("ipop", "solve.QP", "Matlab"))
}
\arguments{
\item{alphaf}{is alpha factors' name.}

\item{constr}{is optimization constraint,\bold{IndSty} means industry and style neutral,
\bold{IndStyTE} means besides industry and style neutral,tracking error also required.}

\item{Fcov}{is the covariance matrix.}

\item{bmk}{is the benckmark for optimization,can be NULL.}
}
\value{
a \bold{port} object.
}
\description{
optimize portfolio weight.
}
\examples{


}
\author{
Andrew Dow
}

