% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{regression_result}
\alias{reg.TS}
\alias{reg.TSFR}
\alias{regression_result}
\title{regression_result}
\usage{
reg.TSFR(TSFR, regType = c("lm", "glm"), glm_wgt = c("sqrtFV", "res"),
  sectorAttr = defaultSectorAttr())

reg.TS(TS, factorLists, regType = c("glm", "lm"), glm_wgt = c("sqrtFV",
  "res"), sectorAttr = defaultSectorAttr(), dure = months(1))
}
\arguments{
\item{TSFR}{a TSFR object.}

\item{regType}{the regress type,the default type is "lm".}
}
\value{
reg.TSFR return a list, contains dataframes of frtn, residual, Rsquare, pfp

reg.TS return a list, contains dataframes of TSFR,frtn, residual, Rsquare
}
\description{
Regression to the TSFR data, calculate factor return, residuals, and R squrare, etc.
}
\examples{
reg.TSFR(TSFR)
}
\author{
Ruifei.yin
}

