% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{regression_result}
\alias{regression_result}
\alias{reg.TSFR}
\alias{reg.TS}
\title{regression_result}
\usage{
reg.TSFR(TSFR, regType = c("glm", "lm"), glm_wgt = c("sqrtFV", "res"),
  sectorAttr = defaultSectorAttr(), secRtnOut = FALSE)

reg.TS(TS, FactorLists, dure = months(1), regType = c("glm", "lm"),
  glm_wgt = c("sqrtFV", "res"), sectorAttr = defaultSectorAttr(),
  secRtnOut = FALSE)
}
\arguments{
\item{TSFR}{a \bold{TSFR} object.}

\item{regType}{the regress type,the default type is "glm".}

\item{glm_wgt}{glm's weight data, default value is sqrt of floating market value.}

\item{sectorAttr}{sector attribute.}

\item{secRtnOut}{whether output sector's return,default value is \code{FALSE}.}

\item{TS}{a \bold{TS} object.}

\item{dure}{see example in \code{\link{getTSR}}.}
}
\value{
return a list, contains dataframes of TSFR,frtn, residual and Rsquare.
}
\description{
Regression to the TSFR data, calculate factor return, residuals, and R squrare, etc.
}
\examples{
RebDates <- getRebDates(as.Date('2014-01-31'),as.Date('2016-09-30'))
TS <- getTS(RebDates,indexID = 'EI000985')
factorIDs <- c("F000002","F000006","F000008")
FactorLists <- buildFactorLists_lcfs(factorIDs,factorRefine=refinePar_default("robust"))
FactorLists <- buildFactorLists(
  buildFactorList(factorFun="gf.NP_YOY",
                  factorPar=list(),
                  factorDir=1),
  factorRefine=refinePar_default("robust"))
FactorLists <- c(tmp,FactorLists)
re <- reg.TS(TS)
re <- reg.TS(TS,FactorLists)
----------------------------------------------------------
TSF <- getMultiFactor(TS,FactorLists)
TSFR <- getTSR(TSF)
re <- reg.TSFR(TSFR)
}
\author{
Ruifei.yin
}
