% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct05_regressionBacktest.R
\name{regression_result_summary}
\alias{MC.chart.regCorr}
\alias{chart.reg.fRtnBar}
\alias{chart.reg.fRtnWealthIndex}
\alias{chart.reg.rsquare}
\alias{regression_result_summary}
\alias{table.reg.fRtn}
\alias{table.reg.rsquare}
\title{regression_result_summary}
\usage{
table.reg.rsquare(reg_results)

table.reg.fRtn(reg_results)

chart.reg.fRtnWealthIndex(reg_results, facet = FALSE)

chart.reg.fRtnBar(reg_results)

chart.reg.rsquare(reg_results)

MC.chart.regCorr(reg_results)
}
\arguments{
\item{reg_results}{is regression_result}

\item{factet}{whether to plot wealth index of factor's return in one graph, the default value is FALSE.}
}
\description{
summary of regression result, such as chart of rsquare and factor return,etc.
}
\seealso{
\link{reg.TSFR}
}

