% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFactorModel-package.R
\docType{package}
\name{RFactorModel}
\alias{RFactorModel}
\alias{factormodel}
\alias{package-factormodel}
\alias{RFactorModel-package}
\title{Building,testing,reporting the factor model.}
\description{
...
}
